% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAgeGroup.R
\name{addAgeGroup}
\alias{addAgeGroup}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addAgeGroup(x, ageGroup = NULL, name = "age_group", compute = TRUE)
}
\arguments{
\item{x}{x}

\item{ageGroup}{...}

\item{name}{...}

\item{compute}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
