% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAttributes.R
\name{addAttributes}
\alias{addAttributes}
\title{Get attributes from one cohort to another}
\usage{
addAttributes(newcohort, oldcohort)
}
\arguments{
\item{newcohort}{cohort to which to attach the attributes}

\item{oldcohort}{cohort from which to get the attributes}
}
\value{
new cohort with added attributes from the other given cohort
}
\description{
Get attributes from one cohort to another
}
