% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBirthDay.R
\name{addBirthday}
\alias{addBirthday}
\title{Add the birth day of an individual to a table}
\usage{
addBirthday(
  x,
  birthday = 0,
  birthdayName = "birthday",
  ageMissingMonth = 1L,
  ageMissingDay = 1L,
  ageImposeMonth = FALSE,
  ageImposeDay = FALSE,
  ageUnit = "years",
  name = NULL
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{birthday}{Day of birth to add.}

\item{birthdayName}{Name of the birthday column to add.}

\item{ageMissingMonth}{Month of the year assigned when month of birth is
missing.}

\item{ageMissingDay}{Day of the month assigned when day of birth is missing.}

\item{ageImposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}

\item{ageImposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}

\item{ageUnit}{Unit in which to express age: \code{"years"}, \code{"months"}, or
\code{"days"}.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\value{
The table with a new column containing the birth day.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function accounts for leap years and corrects the invalid dates to the
next valid date.
}
\examples{
\donttest{
library(PatientProfiles)
library(dplyr)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addBirthday() |>
  glimpse()

cdm$cohort1 |>
  addBirthday(birthday = 5, birthdayName = "bithday_5th") |>
  glimpse()
}

}
