% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBirthDay.R
\name{addBirthdayQuery}
\alias{addBirthdayQuery}
\title{Add the birth day of an individual to a table}
\usage{
addBirthdayQuery(
  x,
  birthdayName = "birthday",
  birthday = 0,
  ageMissingMonth = 1,
  ageMissingDay = 1,
  ageImposeMonth = FALSE,
  ageImposeDay = FALSE,
  ageUnit = "years"
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{birthdayName}{Name of the birthday column to add.}

\item{birthday}{Day of birth to add.}

\item{ageMissingMonth}{Month of the year assigned when month of birth is
missing.}

\item{ageMissingDay}{Day of the month assigned when day of birth is missing.}

\item{ageImposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}

\item{ageImposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}

\item{ageUnit}{Unit in which to express age: \code{"years"}, \code{"months"}, or
\code{"days"}.}
}
\value{
The table with a query that add the new column containing the birth
day.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Same as \code{addBirthday()}, except query is not computed to a table.

The function accounts for leap years and corrects the invalid dates to the
next valid date.
}
\examples{
\donttest{
library(PatientProfiles)
library(dplyr)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addBirthdayQuery() |>
  glimpse()

cdm$cohort1 |>
  addBirthdayQuery(birthday = 5) |>
  glimpse()
}

}
