% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addCdmName}
\alias{addCdmName}
\title{Add cdm name}
\usage{
addCdmName(table, cdm = omopgenerics::cdmReference(table))
}
\arguments{
\item{table}{A table to process.}

\item{cdm}{A \code{cdm_reference} object.}
}
\value{
Table with an extra column with the cdm names
}
\description{
Add cdm name
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addCdmName()
}

}
