% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCohortIntersect.R
\name{addCohortIntersectDate}
\alias{addCohortIntersectDate}
\title{Date of cohorts that are present in a certain window}
\usage{
addCohortIntersectDate(
  x,
  targetCohortTable,
  targetCohortId = NULL,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  targetDate = "cohort_start_date",
  order = "first",
  window = c(0, Inf),
  nameStyle = "{cohort_name}_{window_name}",
  name = NULL
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{targetCohortTable}{Name of the cohort table to intersect with.}

\item{targetCohortId}{Cohort definition IDs to include from
\code{targetCohortTable}. If \code{NULL}, all cohorts are included.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}

\item{censorDate}{Date or name of a date column in \code{x} on which to censor
follow-up. If \code{NULL}, no censoring is applied.}

\item{targetDate}{Name or names of date columns in the target tables to use
for the intersection.}

\item{order}{Which record to use when multiple records occur in a window:
\code{"first"} or \code{"last"}.}

\item{window}{Window or windows of time relative to \code{indexDate} to consider.}

\item{nameStyle}{Naming pattern for the added column or columns. It should
include the required formatting variables. If more than one \code{tableName} is
provided, it must include \code{{table_name}}.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\value{
x along with additional columns for each cohort of interest.
}
\description{
Date of cohorts that are present in a certain window
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addCohortIntersectDate(targetCohortTable = "cohort2")

}

}
