% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addCohortName}
\alias{addCohortName}
\title{Add cohort name for each cohort_definition_id}
\usage{
addCohortName(cohort)
}
\arguments{
\item{cohort}{A \code{cohort_table} object.}
}
\value{
cohort with an extra column with the cohort names
}
\description{
Add cohort name for each cohort_definition_id
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")
cdm$cohort1 |>
  addCohortName()
}

}
