% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addConceptIntersect.R
\name{addConceptIntersectDays}
\alias{addConceptIntersectDays}
\title{It creates column to indicate the days of difference from an index date to a
concept}
\usage{
addConceptIntersectDays(
  x,
  conceptSet,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = list(c(0, Inf)),
  targetDate = "event_start_date",
  order = "first",
  inObservation = TRUE,
  nameStyle = "{concept_name}_{window_name}",
  name = NULL,
  type = "numeric"
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{conceptSet}{A named list of concept sets.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}

\item{censorDate}{Date or name of a date column in \code{x} on which to censor
follow-up. If \code{NULL}, no censoring is applied.}

\item{window}{Window or windows of time relative to \code{indexDate} to consider.}

\item{targetDate}{Name or names of date columns in the target tables to use
for the intersection.}

\item{order}{Which record to use when multiple records occur in a window:
\code{"first"} or \code{"last"}.}

\item{inObservation}{If \code{TRUE}, only records that occur during an observation
period are considered.}

\item{nameStyle}{Naming pattern for the added column or columns. It should
include the required formatting variables. If more than one \code{tableName} is
provided, it must include \code{{table_name}}.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}

\item{type}{Type of the created column(s). Counts, days, age, and observation
durations can be \code{"numeric"} or \code{"integer"}. Flag columns can also be
\code{"logical"}. Field columns can use \code{"auto"} to preserve the source type, or
can be converted to \code{"numeric"}, \code{"integer"}, \code{"logical"}, or \code{"character"}.}
}
\value{
table with added columns with overlap information
}
\description{
It creates column to indicate the days of difference from an index date to a
concept
}
\examples{
\donttest{
library(PatientProfiles)
library(omopgenerics, warn.conflicts = TRUE)
library(dplyr, warn.conflicts = TRUE)

cdm <- mockPatientProfiles(source = "duckdb")

concept <- tibble(
  concept_id = c(1125315),
  domain_id = "Drug",
  vocabulary_id = NA_character_,
  concept_class_id = "Ingredient",
  standard_concept = "S",
  concept_code = NA_character_,
  valid_start_date = as.Date("1900-01-01"),
  valid_end_date = as.Date("2099-01-01"),
  invalid_reason = NA_character_
) |>
  mutate(concept_name = paste0("concept: ", .data$concept_id))
cdm <- insertTable(cdm, "concept", concept)

cdm$cohort1 |>
  addConceptIntersectDays(conceptSet = list("acetaminophen" = 1125315))

}

}
