% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addDateOfBirth}
\alias{addDateOfBirth}
\title{Add a column with the individual birth date}
\usage{
addDateOfBirth(
  x,
  dateOfBirthName = "date_of_birth",
  missingDay = 1,
  missingMonth = 1,
  imposeDay = FALSE,
  imposeMonth = FALSE,
  name = NULL
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{dateOfBirthName}{Name of the date-of-birth column to add.}

\item{missingDay}{Day of the month assigned when day of birth is missing.}

\item{missingMonth}{Month of the year assigned when month of birth is missing.}

\item{imposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}

\item{imposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\value{
The function returns the table x with an extra column that contains
the date of birth.
}
\description{
Add a column with the individual birth date
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addDateOfBirth()

}
}
