% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographicsQuery.R
\name{addDateOfBirthQuery}
\alias{addDateOfBirthQuery}
\title{Query to add a column with the individual birth date}
\usage{
addDateOfBirthQuery(
  x,
  dateOfBirthName = "date_of_birth",
  missingDay = 1,
  missingMonth = 1,
  imposeDay = FALSE,
  imposeMonth = FALSE
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{dateOfBirthName}{Name of the date-of-birth column to add.}

\item{missingDay}{Day of the month assigned when day of birth is missing.}

\item{missingMonth}{Month of the year assigned when month of birth is missing.}

\item{imposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}

\item{imposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}
}
\value{
The function returns the table x with an extra column that contains
the date of birth.
}
\description{
Same as \code{addDateOfBirth()}, except query is not computed to a table.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addDateOfBirthQuery()

}
}
