% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDeath.R
\name{addDeathDate}
\alias{addDeathDate}
\title{Add date of death for individuals. Only death within the same observation
period than \code{indexDate} will be observed.}
\usage{
addDeathDate(
  x,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = c(0, Inf),
  deathDateName = "date_of_death",
  name = NULL
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}

\item{censorDate}{Date or name of a date column in \code{x} on which to censor
follow-up. If \code{NULL}, no censoring is applied.}

\item{window}{Window or windows of time relative to \code{indexDate} to consider.}

\item{deathDateName}{name of the new column to be added.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\value{
table x with the added column with death information added.
}
\description{
Add date of death for individuals. Only death within the same observation
period than \code{indexDate} will be observed.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addDeathDate()

}

}
