% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addDemographics}
\alias{addDemographics}
\title{Compute demographic characteristics at a certain date}
\usage{
addDemographics(
  x,
  indexDate = "cohort_start_date",
  age = TRUE,
  ageName = "age",
  ageMissingMonth = 1,
  ageMissingDay = 1,
  ageImposeMonth = FALSE,
  ageImposeDay = FALSE,
  ageUnit = "years",
  ageGroup = NULL,
  missingAgeGroupValue = "None",
  sex = TRUE,
  sexName = "sex",
  missingSexValue = "None",
  priorObservation = TRUE,
  priorObservationName = "prior_observation",
  priorObservationType = "days",
  futureObservation = TRUE,
  futureObservationName = "future_observation",
  futureObservationType = "days",
  dateOfBirth = FALSE,
  dateOfBirthName = "date_of_birth",
  name = NULL,
  type = "numeric"
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}

\item{age}{If \code{TRUE}, age is calculated relative to \code{indexDate}.}

\item{ageName}{Name of the age column to add.}

\item{ageMissingMonth}{Month of the year assigned when month of birth is
missing.}

\item{ageMissingDay}{Day of the month assigned when day of birth is missing.}

\item{ageImposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}

\item{ageImposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}

\item{ageUnit}{Unit in which to express age: \code{"years"}, \code{"months"}, or
\code{"days"}.}

\item{ageGroup}{If not \code{NULL}, a list of age-group vectors.}

\item{missingAgeGroupValue}{Value to use when age is missing.}

\item{sex}{If \code{TRUE}, sex is identified.}

\item{sexName}{Name of the sex column to add.}

\item{missingSexValue}{Value to use when sex is missing.}

\item{priorObservation}{If \code{TRUE}, the time between the start of the current
observation period and \code{indexDate} is calculated.}

\item{priorObservationName}{Name of the prior-observation column to add.}

\item{priorObservationType}{Whether to return a \code{"date"} or a number of
\code{"days"}.}

\item{futureObservation}{If \code{TRUE}, the time between \code{indexDate} and the end
of the current observation period is calculated.}

\item{futureObservationName}{Name of the future-observation column to add.}

\item{futureObservationType}{Whether to return a \code{"date"} or a number of
\code{"days"}.}

\item{dateOfBirth}{If \code{TRUE}, date of birth is returned.}

\item{dateOfBirthName}{Name of the date-of-birth column to add.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}

\item{type}{Type of the created column(s). Counts, days, age, and observation
durations can be \code{"numeric"} or \code{"integer"}. Flag columns can also be
\code{"logical"}. Field columns can use \code{"auto"} to preserve the source type, or
can be converted to \code{"numeric"}, \code{"integer"}, \code{"logical"}, or \code{"character"}.}
}
\value{
cohort table with the added demographic information columns.
}
\description{
Compute demographic characteristics at a certain date
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addDemographics()

cdm$cohort1 |>
  addDemographics(indexDate = as.Date("2010-01-01"))

}

}
