% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFollowUp.R
\name{addFollowUp}
\alias{addFollowUp}
\title{It adds a column to a cohort table with the number of days of follow up}
\usage{
addFollowUp(
  x,
  start_date = "cohort_start_date",
  end_date = "cohort_end_date",
  name = "follow_up",
  compute = TRUE
)
}
\arguments{
\item{x}{cohort table in which to add follow up of individuals}

\item{start_date}{name of the column containing the start dates}

\item{end_date}{name of the column containing the end dates}

\item{name}{name of the column to hold the result of the enquiry:
number of days of follow up}

\item{tablePrefix}{Whether resultant table will rename. By default: NULL}
}
\value{
cohort table with the added column with follow up days
}
\description{
It adds a column to a cohort table with the number of days of follow up
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
cdm$cohort \%>\% addFollowUp(cdm)
}

}
