% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseLargeScaleCharacteristics.R
\name{addLargeScaleCharacteristics}
\alias{addLargeScaleCharacteristics}
\title{This function is used to add columns with the large scale characteristics of
a cohort table.}
\usage{
addLargeScaleCharacteristics(
  cohort,
  window = list(c(0, Inf)),
  eventInWindow = NULL,
  episodeInWindow = NULL,
  includeSource = FALSE,
  minimumFrequency = 0.005
)
}
\arguments{
\item{cohort}{The cohort to characterise.}

\item{window}{Temporal windows that we want to characterize.}

\item{eventInWindow}{Tables to characterise the events in the window.}

\item{episodeInWindow}{Tables to characterise the episodes in the window.}

\item{includeSource}{Whether to include source concepts.}

\item{minimumFrequency}{Minimum frequency covariates to report.}
}
\value{
The output of this function is the cohort with the new created
columns
}
\description{
This function is used to add columns with the large scale characteristics of
a cohort table.
}
