% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNumberEvent.R
\name{addNumberEvent}
\alias{addNumberEvent}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addNumberEvent(
  x,
  cdm,
  eventTableName,
  filter = NULL,
  window = c(NA, NA),
  name = "number_event",
  eventAt = "cohort_start_date",
  eventDate = "cohort_start_date",
  compute = TRUE
)
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{eventTableName}{...}

\item{filter}{...}

\item{window}{...}

\item{name}{...}

\item{eventAt}{...}

\item{eventDate}{...}

\item{compute}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
