% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNumberVisits.R
\name{addNumberVisits}
\alias{addNumberVisits}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addNumberVisits(x, cdm, window = c(-365, 0))
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{window}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
