% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addObservationPeriod.R
\name{addObservationPeriod}
\alias{addObservationPeriod}
\title{Add a column to the current tibble with the observation period from observation_period table}
\usage{
addObservationPeriod(
  x,
  cdm,
  observationAt = "cohort_start_date",
  column = c("observation_period_start_date", "observation_period_end_date"),
  name = NULL,
  tablePrefix = NULL
)
}
\arguments{
\item{x}{cohort table to which add prior history to}

\item{cdm}{object containing the person table}

\item{observationAt}{name of the date field to use as date in table x}

\item{column}{name of the observational start date and end date in the observation_period table}

\item{name}{name of the columns added for observation start
date and end date in form of "name for start date","name for end date"}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\description{
Add a column to the current tibble with the observation period from observation_period table
}
\examples{
\dontrun{
#create mock tables for testing
cohort1 <- tibble::tibble(
 cohort_definition_id = c("1", "1", "1"),
 subject_id = c("1", "2", "3"),
 cohort_start_date = c(
   as.Date("2010-03-03"),
   as.Date("2010-03-01"),
   as.Date("2010-02-01")
 ),
 cohort_end_date = c(
   as.Date("2015-01-01"),
   as.Date("2013-01-01"),
   as.Date("2013-01-01")
 )
)

obs_1 <- tibble::tibble(
 observation_period_id = c("1", "2", "3"),
 person_id = c("1", "2", "3"),
 observation_period_start_date = c(
   as.Date("2010-02-03"),
   as.Date("2010-02-01"),
   as.Date("2010-01-01")
 ),
 observation_period_end_date = c(
   as.Date("2014-01-01"),
   as.Date("2012-01-01"),
   as.Date("2012-01-01")
 )
)

cdm <-
 mockCohortProfiles(
   seed = 1,
   cohort1 = cohort1,
   observation_period = obs_1

 )

result <- cdm$cohort1 \%>\% addPriorHistory(cdm) \%>\% dplyr::collect()
}
}
