% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addObservationPeriodId.R
\name{addObservationPeriodId}
\alias{addObservationPeriodId}
\title{Add the ordinal number of the observation period associated that a given date
is in.}
\usage{
addObservationPeriodId(
  x,
  indexDate = "cohort_start_date",
  nameObservationPeriodId = "observation_period_id",
  name = NULL
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}

\item{nameObservationPeriodId}{Name of the observation-period ID column to
add.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\value{
Table with the current observation period id added.
}
\description{
Add the ordinal number of the observation period associated that a given date
is in.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addObservationPeriodId()

}

}
