% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addObservationPeriodId.R
\name{addObservationPeriodIdQuery}
\alias{addObservationPeriodIdQuery}
\title{Add the ordinal number of the observation period associated that a given date
is in. Result is not computed, only query is added.}
\usage{
addObservationPeriodIdQuery(
  x,
  indexDate = "cohort_start_date",
  nameObservationPeriodId = "observation_period_id"
)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}

\item{nameObservationPeriodId}{Name of the observation-period ID column to
add.}
}
\value{
Table with the current observation period id added.
}
\description{
Add the ordinal number of the observation period associated that a given date
is in. Result is not computed, only query is added.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addObservationPeriodIdQuery()

}

}
