% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPriorHistory.R
\name{addPriorHistory}
\alias{addPriorHistory}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addPriorHistory(x, cdm, priorHistoryAt = "cohort_start_date", compute = TRUE)
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{priorHistoryAt}{...}

\item{compute}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
