% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addSex}
\alias{addSex}
\title{Compute the sex of the individuals}
\usage{
addSex(x, sexName = "sex", missingSexValue = "None", name = NULL)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{sexName}{Name of the sex column to add.}

\item{missingSexValue}{Value to use when sex is missing.}

\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\value{
table x with the added column with sex information.
}
\description{
Compute the sex of the individuals
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addSex()

}

}
