% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographicsQuery.R
\name{addSexQuery}
\alias{addSexQuery}
\title{Query to add the sex of the individuals}
\usage{
addSexQuery(x, sexName = "sex", missingSexValue = "None")
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{sexName}{Name of the sex column to add.}

\item{missingSexValue}{Value to use when sex is missing.}
}
\value{
table x with the added column with sex information.
}
\description{
Same as \code{addSex()}, except query is not computed to a table.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles(source = "duckdb")

cdm$cohort1 |>
  addSexQuery()

}

}
