% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTableIntersect.R
\name{addTableIntersect}
\alias{addTableIntersect}
\title{It creates columns to indicate overlap information with table in the cdm}
\usage{
addTableIntersect(
  x,
  cdm,
  tableName,
  value = c("number", "binary"),
  window = c(0, Inf),
  filter = NULL,
  order = "first",
  name = "{value}_{tableName}_{window}",
  compute = TRUE
)
}
\arguments{
\item{x}{table containing the individuals for which the overlap indicator
to be attached as extra columns}

\item{cdm}{cdm containing the tables}

\item{tableName}{name of the table that we want to check for overlap}

\item{value}{value of interest to add: it can be number, binary, or a column
of cdm[[tableName]]}

\item{window}{window to consider events of, from date of reference in table x
to date of event at event table}

\item{filter}{filter for the cdm[[tableName]] if provided}

\item{order}{last or first value to use, if not number or binary}

\item{name}{naming of the added column}

\item{compute}{compute/not compute}
}
\value{
table x with added overlap information
}
\description{
It creates columns to indicate overlap information with table in the cdm
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
cdm$cohort2 \%>\% addTableIntersect(cdm, tableName = "drug_exposure",
value = "number", window = c(NA,-1), filter= drug_concept_id == 2)
# This will add a column to the individuals in cohort2 with the number of
drug exposures of drug with drug_concept_id 2, previous to the index event
of the occurrence
}

}
