% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageDoc}
\alias{ageDoc}
\title{Helper for consistent documentation of \code{age}.}
\arguments{
\item{age}{If \code{TRUE}, age is calculated relative to \code{indexDate}.}
}
\description{
Helper for consistent documentation of \code{age}.
}
\keyword{internal}
