% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageGroupDoc}
\alias{ageGroupDoc}
\title{Helper for consistent documentation of \code{ageGroup}.}
\arguments{
\item{ageGroup}{If not \code{NULL}, a list of age-group vectors.}
}
\description{
Helper for consistent documentation of \code{ageGroup}.
}
\keyword{internal}
