% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageImposeDayDoc}
\alias{ageImposeDayDoc}
\title{Helper for consistent documentation of \code{ageImposeDay}.}
\arguments{
\item{ageImposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}
}
\description{
Helper for consistent documentation of \code{ageImposeDay}.
}
\keyword{internal}
