% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageImposeMonthDoc}
\alias{ageImposeMonthDoc}
\title{Helper for consistent documentation of \code{ageImposeMonth}.}
\arguments{
\item{ageImposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}
}
\description{
Helper for consistent documentation of \code{ageImposeMonth}.
}
\keyword{internal}
