% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageMissingDayDoc}
\alias{ageMissingDayDoc}
\title{Helper for consistent documentation of \code{ageMissingDay}.}
\arguments{
\item{ageMissingDay}{Day of the month assigned when day of birth is missing.}
}
\description{
Helper for consistent documentation of \code{ageMissingDay}.
}
\keyword{internal}
