% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageMissingMonthDoc}
\alias{ageMissingMonthDoc}
\title{Helper for consistent documentation of \code{ageMissingMonth}.}
\arguments{
\item{ageMissingMonth}{Month of the year assigned when month of birth is
missing.}
}
\description{
Helper for consistent documentation of \code{ageMissingMonth}.
}
\keyword{internal}
