% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageNameDoc}
\alias{ageNameDoc}
\title{Helper for consistent documentation of \code{ageName}.}
\arguments{
\item{ageName}{Name of the age column to add.}
}
\description{
Helper for consistent documentation of \code{ageName}.
}
\keyword{internal}
