% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ageUnitDoc}
\alias{ageUnitDoc}
\title{Helper for consistent documentation of \code{ageUnit}.}
\arguments{
\item{ageUnit}{Unit in which to express age: \code{"years"}, \code{"months"}, or
\code{"days"}.}
}
\description{
Helper for consistent documentation of \code{ageUnit}.
}
\keyword{internal}
