% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{allowDuplicatesDoc}
\alias{allowDuplicatesDoc}
\title{Helper for consistent documentation of \code{allowDuplicates}.}
\arguments{
\item{allowDuplicates}{Whether to allow multiple records for the same person,
target, and date. If \code{TRUE}, multiple values are collapsed into a
semicolon-separated character value; otherwise, duplicates result in an
error.}
}
\description{
Helper for consistent documentation of \code{allowDuplicates}.
}
\keyword{internal}
