% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{availableEstimates}
\alias{availableEstimates}
\title{Show the available estimates that can be used for the different variable_type
supported.}
\usage{
availableEstimates(variableType = NULL, fullQuantiles = FALSE)
}
\arguments{
\item{variableType}{A set of variable types.}

\item{fullQuantiles}{Whether to display the exact quantiles that can be
computed or only the qXX to summarise all of them.}
}
\value{
A tibble with the available estimates.
}
\description{
Show the available estimates that can be used for the different variable_type
supported.
}
\examples{
\donttest{
library(PatientProfiles)

availableEstimates()
availableEstimates("numeric")
availableEstimates(c("numeric", "categorical"))
}

}
