% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{binaryFormats}
\alias{binaryFormats}
\title{Available functions and formats for binary variables}
\usage{
binaryFormats()
}
\value{
Tibble with the available binary format keys
}
\description{
Available functions and formats for binary variables
}
\examples{
library(PatientProfiles)
binaryFormats()

}
