% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{birthdayDoc}
\alias{birthdayDoc}
\title{Helper for consistent documentation of \code{birthday}.}
\arguments{
\item{birthday}{Day of birth to add.}
}
\description{
Helper for consistent documentation of \code{birthday}.
}
\keyword{internal}
