% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{birthdayNameDoc}
\alias{birthdayNameDoc}
\title{Helper for consistent documentation of \code{birthdayName}.}
\arguments{
\item{birthdayName}{Name of the birthday column to add.}
}
\description{
Helper for consistent documentation of \code{birthdayName}.
}
\keyword{internal}
