% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{categoricalFormats}
\alias{categoricalFormats}
\title{Available functions and formats for categorical variables}
\usage{
categoricalFormats()
}
\value{
Tibble with the available categorical format keys
}
\description{
Available functions and formats for categorical variables
}
\examples{
library(PatientProfiles)
categoricalFormats()

}
