% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{censorDateDoc}
\alias{censorDateDoc}
\title{Helper for consistent documentation of \code{censorDate}.}
\arguments{
\item{censorDate}{Date or name of a date column in \code{x} on which to censor
follow-up. If \code{NULL}, no censoring is applied.}
}
\description{
Helper for consistent documentation of \code{censorDate}.
}
\keyword{internal}
