% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortProfiles-package.R
\docType{package}
\name{CohortProfiles-package}
\alias{CohortProfiles}
\alias{CohortProfiles-package}
\title{CohortProfiles: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk}

Authors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk}
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk}
  \item Kim Lopez-Guell \email{kim.lopez@spc.ox.ac.uk}
}

}
\keyword{internal}
