% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{completeIntervalDoc}
\alias{completeIntervalDoc}
\title{Helper for consistent documentation of \code{completeInterval}.}
\arguments{
\item{completeInterval}{If \code{TRUE}, individuals must be observed for the full
requested interval.}
}
\description{
Helper for consistent documentation of \code{completeInterval}.
}
\keyword{internal}
