% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSmd.R
\name{compute_smd}
\alias{compute_smd}
\title{Summarise standardised mean difference
`r lifecycle::badge("deprecated")`}
\usage{
compute_smd(
  data,
  group = "group_level",
  variables = "variable_name",
  weight = NULL
)
}
\arguments{
\item{data}{summarised characteristics data}

\item{group}{the column to use as group to compute SMD, now only supports
two unique values in this column. Default as "group_level"}

\item{variables}{column with variables to compute SMD for. Default as
variable_name}

\item{weight}{the weight column name in your data}
}
\value{
A tibble with SMD computed
}
\description{
Summarise standardised mean difference
`r lifecycle::badge("deprecated")`
}
\examples{
\donttest{
library(PatientProfiles)
library(dplyr)
cdm <- mockPatientProfiles()

result <- summariseCharacteristics(
  cohort = cdm$cohort2, tableIntersect = list(
    tableName = "visit_occurrence", value = "count", window = c(-365, -1)
  )
)
compute_smd(result, group = "group_level")
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
