% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{conceptSetDoc}
\alias{conceptSetDoc}
\title{Helper for consistent documentation of \code{conceptSet}.}
\arguments{
\item{conceptSet}{A named list of concept sets.}
}
\description{
Helper for consistent documentation of \code{conceptSet}.
}
\keyword{internal}
