% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{dateFormats}
\alias{dateFormats}
\title{Available functions and formats for date variables}
\usage{
dateFormats()
}
\value{
Tibble with the available date format keys
}
\description{
Available functions and formats for date variables
}
\examples{
library(PatientProfiles)
dateFormats()

}
