% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{dateOfBirthDoc}
\alias{dateOfBirthDoc}
\title{Helper for consistent documentation of \code{dateOfBirth}.}
\arguments{
\item{dateOfBirth}{If \code{TRUE}, date of birth is returned.}
}
\description{
Helper for consistent documentation of \code{dateOfBirth}.
}
\keyword{internal}
