% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{dateOfBirthNameDoc}
\alias{dateOfBirthNameDoc}
\title{Helper for consistent documentation of \code{dateOfBirthName}.}
\arguments{
\item{dateOfBirthName}{Name of the date-of-birth column to add.}
}
\description{
Helper for consistent documentation of \code{dateOfBirthName}.
}
\keyword{internal}
