% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortTimeDate.R
\name{dateOfCohort}
\alias{dateOfCohort}
\title{Date of cohorts}
\usage{
dateOfCohort(
  x,
  cdm,
  indexDate = "cohort_start_date",
  targetCohortTable,
  targetCohortId = NULL,
  targetDate = "cohort_start_date",
  order = "first",
  window = c(0, Inf),
  nameStyle = "{value}_{id_name}_{window_name}",
  tablePrefix = NULL
)
}
\arguments{
\item{x}{table containing the individuals for which the time to another
cohort will be added}

\item{cdm}{cdm containing the tables}

\item{indexDate}{Date of interest}

\item{targetCohortTable}{Cohort table to}

\item{targetCohortId}{Cohort IDs of interest from the other cohort table. If
NULL, all cohorts will be used with a time variable added for each
cohort of interest}

\item{targetDate}{Date of interest in the other cohort table. Either
cohort_start_date or cohort_end_date}

\item{order}{date to use if there are multiple records for an
individual during the window of interest. Either first or last.}

\item{window}{Window of time to identify records relative to the indexDate.
Records outside of this time period will be ignored.}

\item{nameStyle}{naming of the added column or columns, should include
required parameters}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\value{
x along with additional columns for each cohort of interest.
}
\description{
Date of cohorts
}
