% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{endDateColumn}
\alias{endDateColumn}
\title{Get the name of the end date column for a certain table in the cdm}
\usage{
endDateColumn(tableName)
}
\arguments{
\item{tableName}{Name of the table}
}
\value{
Name of the end date column in that table
}
\description{
Get the name of the end date column for a certain table in the cdm
}
\examples{
\donttest{
library(PatientProfiles)
endDateColumn("condition_occurrence")
}

}
