% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{fieldDoc}
\alias{fieldDoc}
\title{Helper for consistent documentation of \code{field}.}
\arguments{
\item{field}{Name or names of columns in the target tables to add to \code{x}.}
}
\description{
Helper for consistent documentation of \code{field}.
}
\keyword{internal}
