% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCohortId.R
\name{filterCohortId}
\alias{filterCohortId}
\title{Filter a cohort according to cohort_definition_id column, the result is not
computed into a table. only a query is added. Used usually as internal
functions of other packages.}
\usage{
filterCohortId(cohort, cohortId = NULL)
}
\arguments{
\item{cohort}{A `cohort_table` object.}

\item{cohortId}{A vector with cohort ids.}
}
\value{
A `cohort_table` object.
}
\description{
Filter a cohort according to cohort_definition_id column, the result is not
computed into a table. only a query is added. Used usually as internal
functions of other packages.
}
