% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInObservation.R
\name{filterInObservation}
\alias{filterInObservation}
\title{Filter the rows of a \code{cdm_table} to the ones in observation that \code{indexDate}
is in observation.}
\usage{
filterInObservation(x, indexDate)
}
\arguments{
\item{x}{A table containing individuals in a CDM reference.}

\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}
}
\value{
A \code{cdm_table} that is a subset of the original table.
}
\description{
Filter the rows of a \code{cdm_table} to the ones in observation that \code{indexDate}
is in observation.
}
\examples{
\dontrun{
library(PatientProfiles)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

cdm$condition_occurrence |>
  filterInObservation(indexDate = "condition_start_date")

}

}
