% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummary.R
\name{formatEstimates}
\alias{formatEstimates}
\title{Give format to a summary object.}
\usage{
formatEstimates(
  summaryResult,
  format = c(),
  keepNotFromatted = TRUE,
  decimals = c(count = 0),
  defaultDecimal = 2,
  decimalMark = ".",
  bigMark = ","
)
}
\arguments{
\item{summaryResult}{A SummarisedResult object.}

\item{format}{A list of formats of teh estimate_type.}

\item{keepNotFromatted}{Whether to keep non formatted lines.}

\item{decimals}{Number of decimals to round each estimate_type.}
}
\value{
A formatted summarisedResult.
}
\description{
Give format to a summary object.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort1 \%>\%
  summariseCharacteristics(
    ageGroup = list(c(0, 19), c(20, 39), c(40, 59), c(60, 79), c(80, 150)),
    minCellCount = 1
  ) \%>\%
  formatEstimates(
    format = c(
      "N (\%)" = "count (percentage\%)",
      "N" = "count",
      "median [Q25-Q75]" = "median [q25-q75]"
    ),
    decimals = c(count = 0),
    keepNotFromatted = FALSE
  )
}

}
