% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{futureObservationDoc}
\alias{futureObservationDoc}
\title{Helper for consistent documentation of \code{futureObservation}.}
\arguments{
\item{futureObservation}{If \code{TRUE}, the time between \code{indexDate} and the end
of the current observation period is calculated.}
}
\description{
Helper for consistent documentation of \code{futureObservation}.
}
\keyword{internal}
