% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{futureObservationNameDoc}
\alias{futureObservationNameDoc}
\title{Helper for consistent documentation of \code{futureObservationName}.}
\arguments{
\item{futureObservationName}{Name of the future-observation column to add.}
}
\description{
Helper for consistent documentation of \code{futureObservationName}.
}
\keyword{internal}
