% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{futureObservationTypeDoc}
\alias{futureObservationTypeDoc}
\title{Helper for consistent documentation of \code{futureObservationType}.}
\arguments{
\item{futureObservationType}{Whether to return a \code{"date"} or a number of
\code{"days"}.}
}
\description{
Helper for consistent documentation of \code{futureObservationType}.
}
\keyword{internal}
