% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlappingCohort.R
\name{getOverlappingCohort}
\alias{getOverlappingCohort}
\title{This subject if in a certain window the target cohort overlaps with the
overlap cohort}
\usage{
getOverlappingCohort(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  overlapCohortName,
  overlapCohortId = NULL,
  lookbackWindow = 0,
  multipleEvents = FALSE
)
}
\arguments{
\item{cdm}{CDMConnector CDM reference object}

\item{targetCohortName}{name of the cohort that we want to add the
overlapping}

\item{targetCohortId}{id or vector of ids, of the target cohort}

\item{overlapCohortName}{name of the cohort that we want to check if
overlaps with the target cohort}

\item{overlapCohortId}{id or vector of ids, of the cohort we are interested
in}

\item{lookbackWindow}{lookback period window in days, a list consist of two
numbers}

\item{multipleEvents}{if we want to count the number of evenets (TRUE) then
the result is an integer orjust their presence (FALSE) then the result is 1
or 0.}
}
\value{
it add a new column to target cohort for each id in overlap cohort.
The column is numeric and can be 0 if no overlap is observed and 1 if overlap
is observed.
}
\description{
This subject if in a certain window the target cohort overlaps with the
overlap cohort
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(tibble)
cohort <- tibble::tibble(
  cohort_definition_id = c("1", "2", "1", "2", "2"),
  subject_id = c("1", "1", "2", "2", "3"),
  cohort_start_date = c(
    as.Date("2010-01-01"), as.Date("2009-12-01"),
    as.Date("2010-01-01"), as.Date("2009-01-01"),
    as.Date("2010-01-01")
  ), cohort_end_date = c(
    as.Date("2015-01-01"), as.Date("2013-01-01"),
    as.Date("2015-01-01"), as.Date("2009-01-02"),
    as.Date("2015-01-01")
  )
)
cdm <- mockDrugUtilisation(
  cohort = cohort
)
getOverlappingCohort(
  cdm = cdm,
  targetCohortName = "cohort",
  targetCohortId = 1,
  overlapCohortName = "cohort",
  overlapCohortId = c(2, 3),
  lookbackWindow = c(-180, 0)
)
}

}
