% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableOne.R
\name{getTableOne}
\alias{getTableOne}
\title{This function is used to summarise the dose and/or indication over multiple
cohorts.}
\usage{
getTableOne(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  ageGroups = NULL,
  windowVisitOcurrence = NULL,
  covariatesTableName = NULL,
  covariatesSet = NULL,
  covariatesWindow = NULL,
  ...,
  minimumCellCount = 5
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain the 'doseTableName' table and  also 'strataCohortName' and
'indicationList' if specified. It is a compulsory input, no default
value is provided.}

\item{targetCohortName}{target cohort which table one is created for, character}

\item{targetCohortId}{target cohort definition ids aiming to include, can be a list of integers or NULL. When Null, use all ids}

\item{ageGroups}{A list of age groups we are interested in adding count. Each group should contain a vector with min age and max
age. e.g. `list(c(0,10),c(20,30))` Can be NULL. If NULL, do not consider age groups}

\item{windowVisitOcurrence}{A vector of window, using which visit occurrence count will be checked.}

\item{covariatesTableName}{covariatesTableName}

\item{covariatesSet}{covariatesSet}

\item{covariatesWindow}{covariatesWindow}

\item{...}{you can add as many covariates tables that you want following the
pattern: xxxTableName, xxxSet, xxxWindow, where xxxTableName would be the
cohort table name in the cdm, xxxSet the cohortSet and xxxWindow the window
to asses the covariates. xxx will be the name of}

\item{minimumCellCount}{minimum counts due to obscure}
}
\description{
This function is used to summarise the dose and/or indication over multiple
cohorts.
}
