% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummary.R
\name{gtResult}
\alias{gtResult}
\title{Give format to a summary object.}
\usage{
gtResult(
  summaryResult,
  long = list(Variable = c(level = "variable"), Level = c(level = "variable_level"),
    Format = c(level = "format")),
  wide = list(`CDM Name` = c(level = "cdm_name"), Group = c(level = c("group_name",
    "group_level")), Strata = c(level = c("strata_name", "strata_level"))),
  format = c(`N (\%)` = "count (percentage\%)", "median [min; q25 - q75; max]",
    "mean (sd)", "median [q25 - q75]", N = "count"),
  keepNotFromatted = TRUE,
  eliminateUniqueLabels = TRUE,
  decimals = c(default = 0),
  decimalMark = ".",
  bigMark = ",",
  style = list(defaultWidth = "200px", separator = c("Format"))
)
}
\arguments{
\item{summaryResult}{A SummarisedResult object.}

\item{format}{A list of formats of teh estimate_type.}

\item{keepNotFromatted}{Whether to keep non formatted lines.}

\item{decimals}{Number of decimals to round each estimate_type.}

\item{decimalMark}{Decimal mark.}

\item{bigMark}{Big mark.}

\item{settings}{Columns that contain the setting, the name point to the
label.}

\item{variable}{Name of the clumn that contains the variable name and level.}

\item{defaultDecimal}{Number of decimals by default.}
}
\value{
A formatted summarisedResult.
}
\description{
Give format to a summary object.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort1 \%>\%
  summariseCharacteristics(
    ageGroup = list(c(0, 19), c(20, 39), c(40, 59), c(60, 79), c(80, 150)),
    minCellCount = 1
  ) \%>\%
  gtResult(
    format = c(
      "N (\%)" = "count (percentage\%)",
      "N" = "count",
      "median [Q25-Q75]" = "median [q25-q75]"
    ),
    decimals = c(count = 0),
    keepNotFromatted = FALSE
  )
}

}
