% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{imposeDayDoc}
\alias{imposeDayDoc}
\title{Helper for consistent documentation of \code{imposeDay}.}
\arguments{
\item{imposeDay}{If \code{TRUE}, day of birth is treated as missing for all
individuals.}
}
\description{
Helper for consistent documentation of \code{imposeDay}.
}
\keyword{internal}
