% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{imposeMonthDoc}
\alias{imposeMonthDoc}
\title{Helper for consistent documentation of \code{imposeMonth}.}
\arguments{
\item{imposeMonth}{If \code{TRUE}, month of birth is treated as missing for all
individuals.}
}
\description{
Helper for consistent documentation of \code{imposeMonth}.
}
\keyword{internal}
