% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{indexDateDoc}
\alias{indexDateDoc}
\title{Helper for consistent documentation of \code{indexDate}.}
\arguments{
\item{indexDate}{Name of a date column in \code{x}, or a single date to use for
all rows, used as the reference date.}
}
\description{
Helper for consistent documentation of \code{indexDate}.
}
\keyword{internal}
