% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inObservation.R
\name{inObservation}
\alias{inObservation}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
inObservation(x, cdm, compute = TRUE)
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{compute}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
