% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{inObservationDoc}
\alias{inObservationDoc}
\title{Helper for consistent documentation of \code{inObservation}.}
\arguments{
\item{inObservation}{If \code{TRUE}, only records that occur during an observation
period are considered.}
}
\description{
Helper for consistent documentation of \code{inObservation}.
}
\keyword{internal}
