% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{missingAgeGroupValueDoc}
\alias{missingAgeGroupValueDoc}
\title{Helper for consistent documentation of \code{missingAgeGroupValue}.}
\arguments{
\item{missingAgeGroupValue}{Value to use when age is missing.}
}
\description{
Helper for consistent documentation of \code{missingAgeGroupValue}.
}
\keyword{internal}
