% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{missingDayDoc}
\alias{missingDayDoc}
\title{Helper for consistent documentation of \code{missingDay}.}
\arguments{
\item{missingDay}{Day of the month assigned when day of birth is missing.}
}
\description{
Helper for consistent documentation of \code{missingDay}.
}
\keyword{internal}
