% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{missingMonthDoc}
\alias{missingMonthDoc}
\title{Helper for consistent documentation of \code{missingMonth}.}
\arguments{
\item{missingMonth}{Month of the year assigned when month of birth is missing.}
}
\description{
Helper for consistent documentation of \code{missingMonth}.
}
\keyword{internal}
