% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{missingSexValueDoc}
\alias{missingSexValueDoc}
\title{Helper for consistent documentation of \code{missingSexValue}.}
\arguments{
\item{missingSexValue}{Value to use when sex is missing.}
}
\description{
Helper for consistent documentation of \code{missingSexValue}.
}
\keyword{internal}
