% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockPatientProfiles.R
\name{mockDisconnect}
\alias{mockDisconnect}
\title{Function to disconnect from the mock}
\usage{
mockDisconnect(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\description{
Function to disconnect from the mock
}
