% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameDoc}
\alias{nameDoc}
\title{Helper for consistent documentation of \code{name}.}
\arguments{
\item{name}{Name of the new table. If \code{NULL}, a temporary table is returned.}
}
\description{
Helper for consistent documentation of \code{name}.
}
\keyword{internal}
