% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameObservationPeriodIdDoc}
\alias{nameObservationPeriodIdDoc}
\title{Helper for consistent documentation of \code{nameObservationPeriodId}.}
\arguments{
\item{nameObservationPeriodId}{Name of the observation-period ID column to
add.}
}
\description{
Helper for consistent documentation of \code{nameObservationPeriodId}.
}
\keyword{internal}
