% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameStyleDoc}
\alias{nameStyleDoc}
\title{Helper for consistent documentation of \code{nameStyle}.}
\arguments{
\item{nameStyle}{Naming pattern for the added column or columns. It should
include the required formatting variables. If more than one \code{tableName} is
provided, it must include \code{{table_name}}.}
}
\description{
Helper for consistent documentation of \code{nameStyle}.
}
\keyword{internal}
