% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameStyleEventDoc}
\alias{nameStyleEventDoc}
\title{Helper for event-specific documentation of \code{nameStyle}.}
\arguments{
\item{nameStyle}{Naming pattern for the added columns. It must contain
\code{{value}} and can also contain \code{{window_name}}.}
}
\description{
Helper for event-specific documentation of \code{nameStyle}.
}
\keyword{internal}
