% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{numericFormats}
\alias{numericFormats}
\title{Available functions and formats for numeric variables}
\usage{
numericFormats()
}
\value{
Tibble with the available numeric format keys
}
\description{
Available functions and formats for numeric variables
}
\examples{
library(PatientProfiles)
numericFormats()

}
