% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obscureSummary.R
\name{obscureSummary}
\alias{obscureSummary}
\title{This function is used to obscure small counts in population summary tables}
\usage{
obscureSummary(
  result,
  minimumCellCounts = 5,
  globalVariables = c("number_observations", "number_subjects"),
  estimatesToObscure = "count"
)
}
\arguments{
\item{result}{table with the population summary, must have variables
cohort_definition_id, variable, estimate and value}

\item{minimumCellCounts}{number below which the cell results are obscured}

\item{globalVariables}{variables for which whole cohorts are obscured if
their estimatesToObscure are below the minimumCellCount}

\item{estimatesToObscure}{estimates to obscure if below minimumCellCount}
}
\value{
table with the required cells obscured
}
\description{
This function is used to obscure small counts in population summary tables
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
summary_c1 <- getTableOne(cdm,cohort1)
summary_obscured <- obscureSummary(summary_c1)
}


}
