% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummarised.R
\name{optionsTableCharacteristics}
\alias{optionsTableCharacteristics}
\title{Additional arguments for the function tableCharacteristics.}
\usage{
optionsTableCharacteristics()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
tableCharacteristics, and their given default values.
}
\examples{
{
optionsTableCharacteristics()
}

}
