% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummarised.R
\name{optionsTableCohortOverlap}
\alias{optionsTableCohortOverlap}
\title{Additional arguments for the function tableCohortOverlap.}
\usage{
optionsTableCohortOverlap()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
tableCohortOverlap and their given default value.
}
\examples{
{
optionsTableCohortOverlap()
}


}
