% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummarised.R
\name{optionsTableCohortTiming}
\alias{optionsTableCohortTiming}
\title{Additional arguments for the function tableCohortTiming.}
\usage{
optionsTableCohortTiming()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
tableCohortTiming and their given default value.
}
\examples{
{
optionsTableCohortTiming()
}


}
