% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{orderDoc}
\alias{orderDoc}
\title{Helper for consistent documentation of \code{order}.}
\arguments{
\item{order}{Which record to use when multiple records occur in a window:
\code{"first"} or \code{"last"}.}
}
\description{
Helper for consistent documentation of \code{order}.
}
\keyword{internal}
