% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortIntersect}
\alias{plotCohortIntersect}
\title{Plot summariseCohortIntersect output.}
\usage{
plotCohortIntersect(
  data,
  xAxis = "estimate_value",
  yAxis = "variable_name",
  plotStyle = "barplot",
  facetVarX = "variable_name",
  facetVarY = c("group_level", "strata_level"),
  colorVars = "variable_level",
  vertical_x = TRUE
)
}
\arguments{
\item{data}{output of summariseCohortIntersect}

\item{xAxis}{what to plot on x axis, default as variable_name column. Has to be a column in data.}

\item{yAxis}{what to plot on y axis, default as estimate_value column. Has to be a column in data. One of the xAxis or yAxis has to be estimate_value.}

\item{plotStyle}{Now allows boxplot or barplot only.}

\item{facetVarX}{column in data to facet by on horizontal axis}

\item{facetVarY}{column in data to facet by on vertical axis}

\item{colorVars}{column in data to color by.}

\item{vertical_x}{whether to display x axis string vertically.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("deprecated")`
}
\examples{
\donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
results <- summariseCohortIntersect(
  cohort = cdm$cohort1,
  cohortIntersect = list(
    "Medications in the prior year" = list(
      targetCohortTable = "cohort2", value = "flag", window = c(-365, -1)
    )
  )
)
graph <- plotCohortIntersect(results)
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
