% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Summarise cohort timing}
\usage{
plotCohortOverlap(result, percentage = TRUE, subjects = TRUE, facetBy = NULL)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{percentage}{If TRUE the overlap plot will display percentages,
otherwise it will present counts.}

\item{subjects}{If TRUE the overlap plot will display number of subjects,
otherwise it will present number of records.}

\item{facetBy}{names of columns in the result dataframe for faceting the
ggplot object.}
}
\value{
A summarised result
}
\description{
Summarise cohort timing
}
\examples{
\donttest{
cdm_local <- omock::mockCdmReference() |>
  omock::mockPerson(100) |>
  omock::mockObservationPeriod() |>
  omock::mockCohort(numberCohorts = 2)

con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
cdm <- CDMConnector::copy_cdm_to(con = con,
                                 cdm = cdm_local,
                                 schema = "main")

timing <- summariseCohortTiming(cdm$cohort)
}

}
