% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortTiming}
\alias{plotCohortTiming}
\title{Plot cohort_overlap objects.}
\usage{
plotCohortTiming(
  result,
  type = "boxplot",
  cohortNameReference = NULL,
  cohortNameComparator = NULL,
  strataName = "overall",
  strataLevel = "overall",
  cdmName = NULL,
  facetBy = NULL,
  color = "timingLabel",
  timingLabel = "{cohort_name_reference}; {cohort_name_comparator}",
  uniqueCombinations = TRUE
)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{cohortNameReference}{Names of the reference cohorts to include.}

\item{cohortNameComparator}{Names of the comparator cohorts to include.}

\item{strataName}{Names of the strata names to include.}

\item{strataLevel}{Names of the strata levels to include.}

\item{cdmName}{Name of the databases to include.}

\item{facetBy}{Names of columns in the cohort_overlap table for faceting the
ggplot object.}

\item{uniqueCombinations}{If TRUE, only unique combinations of reference and
comparator plots will be plotted.}

\item{variableName}{Name of the variable names to include.}

\item{overlapLabel}{A glue expression to identify each plotted cohort
overlap.}
}
\value{
A ggplot2.
}
\description{
Plot cohort_overlap objects.
}
\examples{
\donttest{
}

}
