% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummariseLargeScaleCharacteristics.R
\name{plotSummariseLargeScaleCharacteristics}
\alias{plotSummariseLargeScaleCharacteristics}
\title{Summarise characteristics of individuals}
\usage{
plotSummariseLargeScaleCharacteristics(
  data,
  cohort_name,
  compare_cohorts = FALSE,
  cohort_2_name = NULL,
  facet_vars = c("estimate_type"),
  color_vars = c("variable_level", "group_level"),
  color_names = NULL,
  options = list()
)
}
\arguments{
\item{data}{output of summariseLargeScaleCharacteristics}

\item{cohort_name}{the cohort name you want to characterise}

\item{compare_cohorts}{Whether to compare cohort or not}

\item{cohort_2_name}{if is not NULL, need to provide cohort_2_name as second cohort to compare to}

\item{facet_vars}{column in data to facet by}

\item{color_vars}{column in data to color by}

\item{color_names}{A vector or pre-selected color}

\item{options}{Other plot options in a list}
}
\value{
A ggplot
}
\description{
Summarise characteristics of individuals
}
