% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotTableIntersect}
\alias{plotTableIntersect}
\title{Plot summariseTableIntersect output.}
\usage{
plotTableIntersect(
  data,
  xAxis = "variable_name",
  yAxis = "estimate_value",
  plotStyle = "boxplot",
  facetVarX = "variable_name",
  facetVarY = c("group_level", "strata_level"),
  colorVars = NULL,
  vertical_x = TRUE
)
}
\arguments{
\item{data}{output of summariseTableIntersect}

\item{xAxis}{what to plot on x axis, default as variable_name column. Has to be a column in data.}

\item{yAxis}{what to plot on y axis, default as estimate_value column. Has to be a column in data. One of the xAxis or yAxis has to be estimate_value.}

\item{plotStyle}{Now allows boxplot or barplot only.}

\item{facetVarX}{column in data to facet by on horizontal axis}

\item{facetVarY}{column in data to facet by on vertical axis}

\item{colorVars}{column in data to color by.}

\item{vertical_x}{whether to display x axis string vertically.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
