% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{priorObservationDoc}
\alias{priorObservationDoc}
\title{Helper for consistent documentation of \code{priorObservation}.}
\arguments{
\item{priorObservation}{If \code{TRUE}, the time between the start of the current
observation period and \code{indexDate} is calculated.}
}
\description{
Helper for consistent documentation of \code{priorObservation}.
}
\keyword{internal}
