% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{priorObservationNameDoc}
\alias{priorObservationNameDoc}
\title{Helper for consistent documentation of \code{priorObservationName}.}
\arguments{
\item{priorObservationName}{Name of the prior-observation column to add.}
}
\description{
Helper for consistent documentation of \code{priorObservationName}.
}
\keyword{internal}
