% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{priorObservationTypeDoc}
\alias{priorObservationTypeDoc}
\title{Helper for consistent documentation of \code{priorObservationType}.}
\arguments{
\item{priorObservationType}{Whether to return a \code{"date"} or a number of
\code{"days"}.}
}
\description{
Helper for consistent documentation of \code{priorObservationType}.
}
\keyword{internal}
