% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-omopgenerics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{suppress}
\alias{settings}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics:settings]{settings()}}, \code{\link[omopgenerics:suppress]{suppress()}}}
}}

