% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{sexDoc}
\alias{sexDoc}
\title{Helper for consistent documentation of \code{sex}.}
\arguments{
\item{sex}{If \code{TRUE}, sex is identified.}
}
\description{
Helper for consistent documentation of \code{sex}.
}
\keyword{internal}
