% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{sexNameDoc}
\alias{sexNameDoc}
\title{Helper for consistent documentation of \code{sexName}.}
\arguments{
\item{sexName}{Name of the sex column to add.}
}
\description{
Helper for consistent documentation of \code{sexName}.
}
\keyword{internal}
