% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{sourceConceptIdColumn}
\alias{sourceConceptIdColumn}
\title{Get the name of the source concept_id column for a certain table in the cdm}
\usage{
sourceConceptIdColumn(tableName)
}
\arguments{
\item{tableName}{Name of the table.}
}
\value{
Name of the source_concept_id column in that table.
}
\description{
Get the name of the source concept_id column for a certain table in the cdm
}
\examples{
\donttest{
library(PatientProfiles)
sourceConceptIdColumn("condition_occurrence")
}

}
