% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{standardConceptIdColumn}
\alias{standardConceptIdColumn}
\title{Get the name of the standard concept_id column for a certain table in the cdm}
\usage{
standardConceptIdColumn(tableName)
}
\arguments{
\item{tableName}{Name of the table.}
}
\value{
Name of the concept_id column in that table.
}
\description{
Get the name of the standard concept_id column for a certain table in the cdm
}
\examples{
\donttest{
library(PatientProfiles)
standardConceptIdColumn("condition_occurrence")
}

}
