% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{startDateColumn}
\alias{startDateColumn}
\title{Get the name of the start date column for a certain table in the cdm}
\usage{
startDateColumn(tableName)
}
\arguments{
\item{tableName}{Name of the table.}
}
\value{
Name of the start date column in that table.
}
\description{
Get the name of the start date column for a certain table in the cdm
}
\examples{
\donttest{
library(PatientProfiles)
startDateColumn("condition_occurrence")
}

}
