% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCharacteristics.R
\name{summariseCohortCounts}
\alias{summariseCohortCounts}
\title{Summarise counts for each different cohort. You can add a list of
stratifications.}
\usage{
summariseCohortCounts(cohort, strata = list())
}
\arguments{
\item{cohort}{A cohort in the cdm.}

\item{strata}{Stratification list.}
}
\value{
A summary of the number of individuals in each cohort and strata.
}
\description{
`r lifecycle::badge("deprecated")`
}
\examples{
\donttest{
cdm <- mockPatientProfiles()

cdm$cohort1 |>
  addSex() |>
  summariseCohortCounts(strata = "sex")
}

}
