% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortIntersect.R
\name{summariseCohortIntersect}
\alias{summariseCohortIntersect}
\title{Summarise cohort intersection information}
\usage{
summariseCohortIntersect(
  cohort,
  targetCohort = list(),
  strata = list(),
  ageGroup = NULL
)
}
\arguments{
\item{cohort}{A cohort in the cdm.}

\item{targetCohort}{The settings for cohort intersection settings.}

\item{strata}{Stratification list.}

\item{ageGroup}{A list of age groups.}
}
\value{
A summary of the cohort intersection informations.
}
\description{
Summarise cohort intersection information
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

summariseCohortIntersect(
  cohort = cdm$cohort1,
  targetCohort = list(
    targetCohortTable = "cohort2", value = "flag", window = c(-365, -1)
  )
)
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
