% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortOverlap.R
\name{summariseCohortOverlap}
\alias{summariseCohortOverlap}
\title{Summarise cohort overlap}
\usage{
summariseCohortOverlap(cohort)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}
}
\value{
A summarised_result
}
\description{
Summarise cohort overlap
}
\examples{
\donttest{
}
}
