% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortTiming.R
\name{summariseCohortTiming}
\alias{summariseCohortTiming}
\title{Summarise cohort timing}
\usage{
summariseCohortTiming(
  cohort,
  restrictToFirstEntry = TRUE,
  timing = c("min", "q25", "median", "q75", "max")
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{restrictToFirstEntry}{If TRUE only an individual's first entry per
cohort will be considered. If FALSE all entries per individual will be
considered}

\item{timing}{Summary statistics for timing.}
}
\value{
A summarised result
}
\description{
Summarise cohort timing
}
\examples{
\donttest{
}

}
