% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCharacteristics.R
\name{summariseConceptIntersect}
\alias{summariseConceptIntersect}
\title{Summarise concept intersect with a cohort_table}
\usage{
summariseConceptIntersect(cohort, conceptIntersect, strata = list())
}
\arguments{
\item{cohort}{A cohort in the cdm}

\item{conceptIntersect}{A list of arguments that uses addConceptIntersect
function to add variables to summarise.}

\item{strata}{Stratification list}
}
\value{
A summary of the concept intersect of the individuals
}
\description{
Summarise concept intersect with a cohort_table
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

summariseConceptIntersect(
  cohort = cdm$cohort1,
  conceptIntersect = list(
    conceptSet = list(a = 1), value = "flag", window = c(-365, -1)
  )
)
}
}
