% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTableIntersect.R
\name{summariseTableIntersect}
\alias{summariseTableIntersect}
\title{Summarise table intersection information}
\usage{
summariseTableIntersect(cohort, tableIntersect = list(), strata = list())
}
\arguments{
\item{cohort}{A cohort in the cdm.}

\item{tableIntersect}{A list of arguments that uses addTableIntersect
function to add variables to summarise.}

\item{strata}{Stratification list.}
}
\value{
A summary of the table intersections.
}
\description{
Summarise table intersection information
}
\examples{
\donttest{
}
}
