% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppressCount.R
\name{suppressCount}
\alias{suppressCount}
\title{Function to suppress counts in summarised objects}
\usage{
suppressCount(
  result,
  minCellCount = 5,
  variable = "estimate",
  estimateType = "count",
  group = c("group_name", "group_level", "strata_name", "strata_level"),
  groupCount = c("number subjects", "number records"),
  linkEstimates = list(count = "percentage")
)
}
\arguments{
\item{result}{Result object}

\item{minCellCount}{Minimum count of records to report results.}

\item{variable}{Columns to suppress counts.}

\item{estimateType}{Types to suppress counts.}

\item{group}{Columns to group by and supress.}

\item{groupCount}{Identifier of the group count.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress counts in summarised objects
}
