% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseResult.R
\name{suppressCounts}
\alias{suppressCounts}
\title{Function to suppress count}
\usage{
suppressCounts(
  result,
  variable = "variable",
  estimate = "estimate",
  group_name = "group_name",
  group_level = "group_level",
  strata_name = "strata_name",
  strata_level = "strata_level",
  minCellCount = 5
)
}
\arguments{
\item{result}{Table including counts, mean, etc, to be suppressed,
with a column varible, to indicate whether it's count/mean/percentage/..
and a column estimate, to show what is the corresponding number.
Default of column names set to the same as variable names, user can self define too.}

\item{minCellCount}{Minimum count of records to report results.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress count
}
