% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFunctions.R
\name{tableCharacteristics}
\alias{tableCharacteristics}
\title{Format a summarised_characteristics object into a visual table.}
\usage{
tableCharacteristics(
  result,
  type = "gt",
  formatEstimateName = c(`N (\%)` = "<count> (<percentage>\%)", N = "<count>",
    `Median [Q25 - Q75]` = "<median> [<q25> - <q75>]", `Mean (SD)` = "<mean> (<sd>)",
    Range = "<min> to <max>"),
  header = c("group"),
  split = c("group", "strata"),
  groupColumn = NULL,
  minCellCount = 5,
  excludeColumns = c("result_id", "result_type", "package_name", "package_version",
    "estimate_type", "additional_name", "additional_level"),
  .options = list()
)
}
\arguments{
\item{result}{A summarised_characteristics object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{minCellCount}{.}

\item{.options}{See optionsTableCharacteristics() for default values.}

\item{splitStrata}{.}

\item{format}{.}

\item{cdmName}{.}

\item{cohortName}{.}

\item{style}{.}
}
\value{
A tibble with a tidy version of the summarised_characteristics
object.
}
\description{
Format a summarised_characteristics object into a visual table.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort1 |>
  summariseCharacteristics() |>
  formatCharacteristics()
}

}
