% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattedTables.R
\name{tableCohortOverlapOptions}
\alias{tableCohortOverlapOptions}
\title{A formatted visualization of cohort_overlap objects.}
\usage{
tableCohortOverlapOptions()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in
tableCohortOverlap and their given default value.
}
\examples{
{
tableCohortOverlapOptions()
}


}
