% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{tableCohortTiming}
\alias{tableCohortTiming}
\title{Format a cohort_timing object into a visual table.}
\usage{
tableCohortTiming(
  result,
  cohortNameReference = NULL,
  cohortNameComparator = NULL,
  cdmName = NULL,
  variableName = c("number records", "number subjects", "diff_days"),
  formatEstimateName = c(N = "<count>", `Mean (SD)` = "<mean> (<sd>)",
    `Median [Q25 - Q75]` = "<median> [<q25> - <q75>]", Range = "<min> - <max>"),
  type = "gt",
  minCellCount = 5,
  .options = list()
)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{cohortNameReference}{Names of the reference cohorts to include.}

\item{cohortNameComparator}{Names of the comparator cohorts to include.}

\item{cdmName}{Name of the databases to include.}

\item{variableName}{Name of the variable names to include.}

\item{formatEstimateName}{Whether to include the number of subjects.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", or "tibble".}

\item{minCellCount}{Counts below which results will be clouded.}

\item{.options}{named list with additional formatting options.
PatientProfiles::optionsTableCohortOverlap() shows allowed arguments and
their default values.}
}
\value{
A formatted table of the cohort_timing summarised object.
}
\description{
Format a cohort_timing object into a visual table.
}
\examples{
\donttest{
cdm_local <- omock::mockCdmReference() |>
  omock::mockPerson(100) |>
  omock::mockObservationPeriod() |>
  omock::mockCohort(numberCohorts = 2)

con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
cdm <- CDMConnector::copy_cdm_to(con = con,
                                 cdm = cdm_local,
                                 schema = "main"
cdm$cohort |>
  summariseCohortOverlap() |>
  tableCohortOverlap()
}

}
