% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{tableNameDoc}
\alias{tableNameDoc}
\title{Helper for consistent documentation of \code{tableName}.}
\arguments{
\item{tableName}{Names of one or more OMOP CDM tables to intersect with.}
}
\description{
Helper for consistent documentation of \code{tableName}.
}
\keyword{internal}
