% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummary.R
\name{tableSummary}
\alias{tableSummary}
\title{Create a gt table from a summary results objects}
\usage{
tableSummary(
  table,
  filterRow = NULL,
  pivotWider = NULL,
  estimateColumn = "estimate",
  hideColumn = NULL
)
}
\arguments{
\item{table}{Long table.}

\item{pivotWider}{List of columns to compare.}

\item{filter}{Filter variables.}

\item{hide}{Columns to hide.}
}
\value{
New table in gt format
}
\description{
Create a gt table from a summary results objects
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  summariseCharacteristics(cdm = cdm, minCellCount = 1) \%>\%
  tableSummary()
}

}
