% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{targetCohortIdDoc}
\alias{targetCohortIdDoc}
\title{Helper for consistent documentation of \code{targetCohortId}.}
\arguments{
\item{targetCohortId}{Cohort definition IDs to include from
\code{targetCohortTable}. If \code{NULL}, all cohorts are included.}
}
\description{
Helper for consistent documentation of \code{targetCohortId}.
}
\keyword{internal}
