% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{targetCohortTableDoc}
\alias{targetCohortTableDoc}
\title{Helper for consistent documentation of \code{targetCohortTable}.}
\arguments{
\item{targetCohortTable}{Name of the cohort table to intersect with.}
}
\description{
Helper for consistent documentation of \code{targetCohortTable}.
}
\keyword{internal}
