% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{targetDateDoc}
\alias{targetDateDoc}
\title{Helper for consistent documentation of \code{targetDate}.}
\arguments{
\item{targetDate}{Name or names of date columns in the target tables to use
for the intersection.}
}
\description{
Helper for consistent documentation of \code{targetDate}.
}
\keyword{internal}
