% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{targetEndDateDoc}
\alias{targetEndDateDoc}
\title{Helper for consistent documentation of \code{targetEndDate}.}
\arguments{
\item{targetEndDate}{Name or names of end-date columns in the target tables
to use for the intersection. If \code{NULL}, the target is treated as a point
event.}
}
\description{
Helper for consistent documentation of \code{targetEndDate}.
}
\keyword{internal}
