% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{targetStartDateDoc}
\alias{targetStartDateDoc}
\title{Helper for consistent documentation of \code{targetStartDate}.}
\arguments{
\item{targetStartDate}{Name or names of start-date columns in the target
tables to use for the intersection.}
}
\description{
Helper for consistent documentation of \code{targetStartDate}.
}
\keyword{internal}
