% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyTables.R
\name{tidyCharacteristics}
\alias{tidyCharacteristics}
\title{A tidy implementation of the summarised_characteristics object.}
\usage{
tidyCharacteristics(result, minCellCount = 5)
}
\arguments{
\item{result}{A summarised_characteristics object.}

\item{minCellCount}{Minimum number of counts to report.}
}
\value{
A tibble with a tidy version of the summarised_characteristics
object.
}
\description{
A tidy implementation of the summarised_characteristics object.
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

cdm$cohort1 |>
  summariseCharacteristics() |>
  tidyCharacteristics()
}

}
