% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{variableTypes}
\alias{variableTypes}
\title{Classify the variables between 5 types: "numeric", "categorical", "logical",
"date", "integer", or NA.}
\usage{
variableTypes(table)
}
\arguments{
\item{table}{A table to process.}
}
\value{
Tibble with the variables type and classification.
}
\description{
Classify the variables between 5 types: "numeric", "categorical", "logical",
"date", "integer", or NA.
}
\examples{
\donttest{
library(PatientProfiles)
library(dplyr, warn.conflicts = TRUE)

x <- tibble(
  person_id = c(1, 2),
  start_date = as.Date(c("2020-05-02", "2021-11-19")),
  asthma = c(0, 1)
)

variableTypes(x)
}

}
