% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAge.R
\name{addAge}
\alias{addAge}
\title{Add a column to the current tibble with the age of the subject_id at a
certain date}
\usage{
addAge(
  x,
  cdm,
  ageAt = "cohort_start_date",
  defaultMonth = 1,
  defaultDay = 1,
  imposeMonth = TRUE,
  imposeDay = TRUE,
  compute = TRUE
)
}
\arguments{
\item{x}{Tibble with the individuals that we want to add the age.}

\item{cdm}{Object that contains a cdm reference. Use CDMConnector to obtain a
cdm reference.}

\item{ageAt}{Variable that points the date to compute the age. By default:
'cohort_start_date'}

\item{defaultMonth}{Month of the year assigned to individuals with missing
month of birth. By default: 1.}

\item{defaultDay}{day of the month assigned to individuals with missing day
of birth. By default: 1.}

\item{imposeMonth}{Weather the month of the date of birth will be considered
as missing for all the individuals. By default: TRUE.}

\item{imposeDay}{Weather the day of the date of birth will be considered as
missing for all the individuals. By default: TRUE.}

\item{compute}{Weather resultant table will be computed as temporal table. By
default: TRUE.}
}
\description{
Add a column to the current tibble with the age of the subject_id at a
certain date
}
