% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAge.R
\name{addAge}
\alias{addAge}
\title{Add a column to the current tibble with the age of the subject_id at a
certain date}
\usage{
addAge(
  x,
  cdm,
  ageAt = "cohort_start_date",
  defaultMonth = 1,
  defaultDay = 1,
  imposeMonth = TRUE,
  imposeDay = TRUE,
  compute = TRUE
)
}
\arguments{
\item{x}{Tibble with the individuals that we want to add the age. Need to be in cdm.}

\item{cdm}{Object that contains a cdm reference. Use CDMConnector to obtain a
cdm reference.}

\item{ageAt}{Variable that points the date to compute the age. By default:
'cohort_start_date'}

\item{defaultMonth}{Month of the year assigned to individuals with missing
month of birth. By default: 1.}

\item{defaultDay}{day of the month assigned to individuals with missing day
of birth. By default: 1.}

\item{imposeMonth}{Whether the month of the date of birth will be considered
as missing for all the individuals. By default: TRUE.}

\item{imposeDay}{Whether the day of the date of birth will be considered as
missing for all the individuals. By default: TRUE.}

\item{compute}{Whether resultant table will be computed as temporal table. By
default: TRUE.}
}
\value{
tibble with the age column added
}
\description{
Add a column to the current tibble with the age of the subject_id at a
certain date
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(tibble)
library(CohortProfiles)
cohort1 <- tibble::tibble(
  cohort_definition_id = c("1", "1", "1"),
  subject_id = c("1", "2", "3"),
  cohort_start_date = c(
    as.Date("2010-01-01"), as.Date("2010-01-01"), as.Date("2010-01-01")
  ),
  cohort_end_date = c(
    as.Date("2015-01-01"), as.Date("2013-01-01"), as.Date("2018-01-01")
  )
)

person <- tibble::tibble(
  person_id = c("1", "2", "3"),
  gender_concept_id = c("8507", "8532", "8507"),
  year_of_birth = c(2000, 1995, NA),
  month_of_birth = c(NA, 07, 08),
  day_of_birth = c(01, 25, 03)
)
cdm <- mockDrugUtilisation(person = person, cohort1 = cohort1)
addAge(x = cdm[["cohort1"]], cdm = cdm)
}
}
