% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAttributes.R
\name{addCdmName}
\alias{addCdmName}
\title{Add cdm name}
\usage{
addCdmName(table, cdm = attr(table, "cdm_reference"))
}
\arguments{
\item{table}{Table in the cdm}

\item{cdm}{A cdm reference object}
}
\value{
Table with an extra column with the cdm names
}
\description{
Add cdm name
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addCdmName()
}

}
