% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDateOfBirth.R
\name{addDateOfBirth}
\alias{addDateOfBirth}
\title{Add a column with the individual birth date}
\usage{
addDateOfBirth(
  x,
  cdm,
  name = "birth_date",
  missingDay = 1,
  misisngMonth = 1,
  imposeDay = FALSE,
  imposeMonth = FALSE
)
}
\arguments{
\item{x}{Table in the cdm that contains 'person_id' or 'subject_id'}

\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con().}

\item{name}{Name of the column to be added with the date of birth}

\item{missingDay}{Day of the individuals with no or imposed day of birth}

\item{imposeDay}{Wether to impose day of birth}

\item{imposeMonth}{Wether to impose month of birth}

\item{missingMonth}{Month of the individuals with no or imposed month of
birth}
}
\value{
The function returns the table x with an extra column that contains
the date of birth
}
\description{
Add a column with the individual birth date
}
\examples{
\donttest{
  library(PatientProfiles)
  db <- DBI::dbConnect(duckdb::duckdb(), CDConnector::eunomia_dir())
  cdm <- mockCdm(db, ...)
  cdm$person \%>\%
    addDateOfBirth(cdm)
  DBI::dbDisconnect(db)
}
}
