% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDeath.R
\name{addDeathDate}
\alias{addDeathDate}
\title{Add date of death for individuals}
\usage{
addDeathDate(
  x,
  indexDate = "cohort_start_date",
  window = c(0, Inf),
  deathDateName = "date_of_death"
)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{indexDate}{Variable in x that contains the date to compute the age.}

\item{window}{window to consider events over.}

\item{deathDateName}{name of the new column to be added.}
}
\value{
table x with the added column with death information added.
}
\description{
Add date of death for individuals
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addDeathDate()
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
