% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDeath.R
\name{addDeathDays}
\alias{addDeathDays}
\title{Add days to death for individuals}
\usage{
addDeathDays(
  x,
  indexDate = "cohort_start_date",
  window = c(0, Inf),
  deathDaysName = "days_to_death"
)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{indexDate}{Variable in x that contains the date to compute the age.}

\item{window}{window to consider events over.}

\item{deathDaysName}{name of the new column to be added.}
}
\value{
table x with the added column with death information added.
}
\description{
Add days to death for individuals
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addDeathDays()
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
