% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDeath.R
\name{addDeathFlag}
\alias{addDeathFlag}
\title{Add flag for death for individuals. Only death within the same observation
period than `indexDate` will be observed.}
\usage{
addDeathFlag(
  x,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  window = c(0, Inf),
  deathFlagName = "death"
)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{indexDate}{Variable in x that contains the window origin.}

\item{censorDate}{Name of a column to stop followup.}

\item{window}{window to consider events over.}

\item{deathFlagName}{name of the new column to be added.}
}
\value{
table x with the added column with death information added.
}
\description{
Add flag for death for individuals. Only death within the same observation
period than `indexDate` will be observed.
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addDeathFlag()
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
