% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addDemographics}
\alias{addDemographics}
\title{It adds all demographics columns to the given cohort table: Age, Sex,
PriorHistory, and ageGroup if desired}
\usage{
addDemographics(
  x,
  cdm,
  indexDate = "cohort_start_date",
  age = TRUE,
  ageDefaultMonth = 1,
  ageDefaultDay = 1,
  ageImposeMonth = TRUE,
  ageImposeDay = TRUE,
  ageGroup = NULL,
  sex = TRUE,
  priorHistory = TRUE,
  tablePrefix = NULL
)
}
\arguments{
\item{x}{cohort table in which to add follow up of individuals}

\item{cdm}{cdm with the person and observation_period tables to get the info
for the individuals in the cohort}

\item{indexDate}{name of the column with the date at which consider
demographic information}

\item{age}{TRUE or FALSE. If TRUE, age will be calculated relative to
indexDate}

\item{ageDefaultMonth}{Month of the year assigned to individuals with missing
month of birth. By default: 1.}

\item{ageDefaultDay}{day of the month assigned to individuals with missing day
of birth. By default: 1.}

\item{ageImposeMonth}{Whether the month of the date of birth will be considered
as missing for all the individuals. By default: TRUE.}

\item{ageImposeDay}{Whether the day of the date of birth will be considered as
missing for all the individuals. By default: TRUE.}

\item{ageGroup}{if not NULL, a list of ageGroup vectors}

\item{sex}{TRUE or FALSE. If TRUE, sex will be identified}

\item{priorHistory}{TRUE or FALSE. If TRUE, days of prior history will
be calculated relative to indexDate}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\value{
cohort table with the added demographic information columns
}
\description{
It adds all demographics columns to the given cohort table: Age, Sex,
PriorHistory, and ageGroup if desired
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
cdm$cohort \%>\% addDemographics(cdm)
}

}
