% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addDemographics}
\alias{addDemographics}
\title{It adds all demographics columns to the given cohort table: Age, Sex,
PriorHistory, and ageGroup if desired}
\usage{
addDemographics(
  x,
  cdm,
  demographicsAt = "cohort_start_date",
  ageGroup = NULL,
  tablePrefix = NULL
)
}
\arguments{
\item{x}{cohort table in which to add follow up of individuals}

\item{cdm}{cdm with the person and observation_period tables to get the info
for the individuals in the cohort}

\item{demographicsAt}{name of the column with the date at which consider
demographic information}

\item{ageGroup}{if not NULL, a list of ageGroup vectors}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\value{
cohort table with the added demographic information columns
}
\description{
It adds all demographics columns to the given cohort table: Age, Sex,
PriorHistory, and ageGroup if desired
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
cdm$cohort \%>\% addDemographics(cdm)
}

}
