% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEvent.R
\name{addEvent}
\alias{addEvent}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addEvent(
  x,
  cdm,
  eventTableName,
  eventId,
  window = c(NA, NA),
  name = "event",
  eventDate = "cohort_start_date",
  eventAt = "cohort_start_date",
  order = "first",
  compute = TRUE
)
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{eventTableName}{...}

\item{eventId}{...}

\item{window}{...}

\item{name}{...}

\item{eventDate}{...}

\item{eventAt}{...}

\item{order}{...}

\item{compute}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
