% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEvent.R
\name{addEvent}
\alias{addEvent}
\title{It adds columns of user defined events for individual in a cohort table}
\usage{
addEvent(
  x,
  cdm,
  eventTableName,
  filter = NULL,
  window = c(NA, NA),
  name = "event",
  eventDate = "cohort_start_date",
  eventAt = "cohort_start_date",
  order = "first",
  compute = TRUE
)
}
\arguments{
\item{x}{table containing the individual for which the event to be attached as extra columns}

\item{cdm}{cdm containing the tables}

\item{eventTableName}{name of the table of the cdm containing the events
of interest}

\item{filter}{condition to apply to the events table for what events to class as event}

\item{window}{window to consider events of, from date of reference in table x
to date of event at event table}

\item{name}{the name of the column containing the event date of the event.}

\item{eventDate}{name of the date field to use as date in the event table}

\item{eventAt}{name of the date field to use as date in table x}

\item{order}{define which event to include first = include earliest event for each individual in x,
last= include last event for each individual in x or all = include all event.}

\item{compute}{if TRUE compute the output}
}
\description{
It adds columns of user defined events for individual in a cohort table
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(tibble)
library(CohortProfiles)
cohort1 <- tibble::tibble(
cohort_definition_id = c("1", "1", "1"),
subject_id = c("1", "2", "3"),
cohort_start_date = c(
 as.Date("2010-03-03"),
 as.Date("2010-03-01"),
 as.Date("2010-02-01")
),
cohort_end_date = c(
 as.Date("2015-01-01"),
 as.Date("2013-01-01"),
 as.Date("2013-01-01")
)
)
cdm <- mockCohortProfiles(seed = 100, cohort1 = cohort1)




addEvent(
   cdm$cohort1,
   cdm,
   "drug_exposure",
   window = c(0, 10000),
   eventDate = "drug_exposure_start_date",
   order = "first"
 )
}
}
