% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addInObservation}
\alias{addInObservation}
\title{Indicate if a certain record is within the observation period}
\usage{
addInObservation(
  x,
  cdm,
  indexDate = "cohort_start_date",
  name = "in_observation",
  temporary = TRUE
)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{cdm}{Object that contains a cdm reference. Use CDMConnector to obtain a
cdm reference.}

\item{indexDate}{Variable in x that contains the date to compute the
observation flag.}

\item{name}{name of the column to hold the result of the enquiry:
1 if the individual is in observation, 0 if not}

\item{temporary}{Whether the resultant table should be temporary or
permanent.}
}
\value{
cohort table with the added binary column assessing inObservation
}
\description{
Indicate if a certain record is within the observation period
}
\examples{
\donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\% addInObservation(cdm)
}

}
