% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInObservation.R
\name{addInObservation}
\alias{addInObservation}
\title{It adds a column to a cohort table indicating whether its individuals are
in observation at the desired time}
\usage{
addInObservation(
  x,
  cdm,
  observationAt = "cohort_start_date",
  name = "in_observation",
  compute = TRUE
)
}
\arguments{
\item{x}{cohort table in which the inObservation command wants to be tested}

\item{cdm}{where the observation_period table is stored}

\item{observationAt}{name of the column with the dates to test the
inObservation command}

\item{name}{name of the column to hold the result of the enquiry:
1 if the individual is in observation, 0 if not}

\item{tablePrefix}{Whether resultant table will rename. By default: NULL}
}
\value{
cohort table with the added binary column assessing inObservation
}
\description{
It adds a column to a cohort table indicating whether its individuals are
in observation at the desired time
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
cdm$cohort \%>\% inObservation(cdm)
}

}
