% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addInObservation}
\alias{addInObservation}
\title{Indicate if a certain record is within the observation period}
\usage{
addInObservation(
  x,
  cdm = lifecycle::deprecated(),
  indexDate = "cohort_start_date",
  window = c(0, 0),
  completeInterval = FALSE,
  name = lifecycle::deprecated(),
  nameStyle = "in_observation"
)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{cdm}{A cdm_reference object.}

\item{indexDate}{Variable in x that contains the date to compute the
observation flag.}

\item{window}{window to consider events of}

\item{completeInterval}{If the individuals are in observation for the full window}

\item{name}{deprecated}

\item{nameStyle}{Name of the new columns to create, it must contain
"{window_name}" if multiple windows are provided.}
}
\value{
cohort table with the added binary column assessing inObservation
}
\description{
Indicate if a certain record is within the observation period
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addInObservation()
}

}
