% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInObservation.R
\name{addInObservation}
\alias{addInObservation}
\title{It adds a column to a cohort table indicating whether its individuals are
in observation at the desired time}
\usage{
addInObservation(
  x,
  cdm,
  indexDate = "cohort_start_date",
  name = "in_observation",
  tablePrefix = NULL
)
}
\arguments{
\item{x}{cohort table in which the inObservation command wants to be tested}

\item{cdm}{where the observation_period table is stored}

\item{indexDate}{name of the column with the dates to test the
inObservation command}

\item{name}{name of the column to hold the result of the enquiry:
1 if the individual is in observation, 0 if not}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\value{
cohort table with the added binary column assessing inObservation
}
\description{
It adds a column to a cohort table indicating whether its individuals are
in observation at the desired time
}
\examples{
\donttest{
library(PatientProfiles)
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\% addInObservation(cdm)
}

}
