% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNumberEvent.R
\name{addNumberEvent}
\alias{addNumberEvent}
\title{It adds a column of counts of desired events per individual to an events
table}
\usage{
addNumberEvent(
  x,
  cdm,
  eventTableName,
  filter = NULL,
  window = c(NA, NA),
  name = "number_event",
  eventAt = "cohort_start_date",
  eventDate = "cohort_start_date",
  compute = TRUE
)
}
\arguments{
\item{x}{table containing the individuals for which the number of events
wants to be counted}

\item{cdm}{containing the events table}

\item{eventTableName}{name of the table of the cdm containing the events
of interest}

\item{filter}{condition to apply to the events table for what events to count}

\item{window}{window to consider events of, from date of reference in table x
to date of event at event table}

\item{name}{name of the column containing the counts per individual}

\item{eventAt}{date of reference in table x}

\item{eventDate}{date to consider in event table}

\item{compute}{whether compute functionality is wanted}
}
\value{
a table with all the individuals in x with a column containing counts
of desired events
}
\description{
It adds a column of counts of desired events per individual to an events
table
}
