% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addObservationPeriodId.R
\name{addObservationPeriodId}
\alias{addObservationPeriodId}
\title{Add the ID associated with current observation period}
\usage{
addObservationPeriodId(x, indexDate = "cohort_start_date", name = NULL)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{indexDate}{Variable in x that contains the date to compute the
observation flag.}

\item{name}{Name of the new table, if NULL a temporary table is returned.}
}
\value{
Table with the current observation period id added.
}
\description{
Add the ID associated with current observation period
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addObservationPeriodId()
mockDisconnect(cdm = cdm)
}

}
