% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addObservationPeriodId.R
\name{addObservationPeriodIdQuery}
\alias{addObservationPeriodIdQuery}
\title{Add the ordinal number of the observation period associated that a given date
is in. Result is not computed, only query is added.}
\usage{
addObservationPeriodIdQuery(
  x,
  indexDate = "cohort_start_date",
  nameObservationPeriodId = "observation_period_id"
)
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{indexDate}{Variable in x that contains the date to compute the
observation flag.}

\item{nameObservationPeriodId}{Name of the new colum.}
}
\value{
Table with the current observation period id added.
}
\description{
Add the ordinal number of the observation period associated that a given date
is in. Result is not computed, only query is added.
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addObservationPeriodIdQuery()
mockDisconnect(cdm = cdm)
}

}
