% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverlap.R
\name{addOverlap}
\alias{addOverlap}
\title{It creates a column to indicate overlaps with the overlapcohort}
\usage{
addOverlap(
  x,
  cdm,
  overlapCohortName,
  overlapCohortId = NULL,
  window = c(NA, NA),
  name = "overlap",
  compute = TRUE
)
}
\arguments{
\item{x}{table containing the individual for which the overlap indicator to be attached as extra columns}

\item{cdm}{cdm containing the tables}

\item{overlapCohortName}{name of the cohort that we want to check if
overlaps with the target cohort}

\item{overlapCohortId}{...}

\item{window}{window to consider events of, from date of reference in table x
to date of event at event table}

\item{name}{name of the new column created for the overlap indicator}

\item{compute}{if TRUE compute the output}
}
\description{
It creates a column to indicate overlaps with the overlapcohort
}
