% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOverlap.R
\name{addOverlap}
\alias{addOverlap}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addOverlap(
  x,
  cdm,
  overlapCohortName,
  filter = NULL,
  window = c(NA, NA),
  name = "overlap"
)
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{overlapCohortName}{...}

\item{filter}{...}

\item{window}{...}

\item{name}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
