% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSex.R
\name{addSex}
\alias{addSex}
\title{It creates a mock database for testing CohortProfiles package}
\usage{
addSex(x, cdm, compute = TRUE)
}
\arguments{
\item{x}{...}

\item{cdm}{...}

\item{compute}{...}
}
\description{
It creates a mock database for testing CohortProfiles package
}
