% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSex.R
\name{addSex}
\alias{addSex}
\title{Adds a column with sex information to the individuals of a table}
\usage{
addSex(x, cdm, name = "sex", compute = TRUE)
}
\arguments{
\item{x}{cohort table to which add Sex}

\item{cdm}{object containing the person table with the sex information
in gender_concept_id column}

\item{name}{name of the new column to be added}

\item{tablePrefix}{Whether resultant table will rename. By default: NULL}
}
\value{
table x with the added column with sex information
}
\description{
Adds a column with sex information to the individuals of a table
}
\examples{
\dontrun{
db <- DBI::dbConnect(" Your database connection here")
cdm <- CDMConnector::cdm_from_con(
  con = db,
  cdm_schema = "cdm schema name"
)
cdm$cohort \%>\% addSex(cdm)
}

}
