% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographics.R
\name{addSex}
\alias{addSex}
\title{Compute the sex of the individuals}
\usage{
addSex(x, cdm = lifecycle::deprecated(), sexName = "sex")
}
\arguments{
\item{x}{Table with individuals in the cdm}

\item{cdm}{A cdm_reference object.}

\item{sexName}{name of the new column to be added}
}
\value{
table x with the added column with sex information
}
\description{
Compute the sex of the individuals
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addSex()
}

}
