% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDemographicsQuery.R
\name{addSexQuery}
\alias{addSexQuery}
\title{Query to add the sex of the individuals}
\usage{
addSexQuery(x, sexName = "sex", missingSexValue = "None")
}
\arguments{
\item{x}{Table with individuals in the cdm.}

\item{sexName}{name of the new column to be added.}

\item{missingSexValue}{Value to include if missing sex.}
}
\value{
table x with the added column with sex information.
}
\description{
Query to add the sex of the individuals
}
\examples{
\donttest{
cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  addSexQuery()
mockDisconnect(cdm = cdm)
}

}
