% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{availableFunctions}
\alias{availableFunctions}
\title{Show the available functions for the 4 classifications of data that are
supported (numeric, date, binary and categorical)}
\usage{
availableFunctions(variableType = NULL)
}
\arguments{
\item{variableType}{A choice between: "numeric", "date", "binary"
or "categorical".}
}
\value{
A tibble with the available functions for a certain variable
classification (or all if NULL)
}
\description{
Show the available functions for the 4 classifications of data that are
supported (numeric, date, binary and categorical)
}
\examples{
\donttest{
library(PatientProfiles)

availableFunctions()
availableFunctions("numeric")
availableFunctions("date")
availableFunctions("binary")
availableFunctions("categorical")
}

}
