% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkPatientProfiles.R
\name{benchmarkPatientProfiles}
\alias{benchmarkPatientProfiles}
\title{Benchmark intersections and demographics functions for a certain source
(cdm).}
\usage{
benchmarkPatientProfiles(cdm, n = 50000, iterations = 1)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{n}{Size of the synthetic cohorts used to benchmark.}

\item{iterations}{Number of iterations to run the benchmark.}
}
\value{
A summarise_result object with the summary statistics.
}
\description{
Benchmark intersections and demographics functions for a certain source
(cdm).
}
