% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtResult.R
\name{gtCharacteristics}
\alias{gtCharacteristics}
\title{Create a gt table from a summary characteristics object}
\usage{
gtCharacteristics(
  summarisedResults,
  pivotWide = c("CDM Name", "Group", "Strata"),
  format = c(`N (\%)` = "count (percentage\%)", "median [min; q25 - q75; max]",
    "mean (sd)", "median [q25 - q75]", N = "count"),
  keepNotFormatted = TRUE,
  eliminateUniqueLabels = TRUE,
  decimals = c(default = 0),
  decimalMark = ".",
  bigMark = ",",
  style = list(defaultWidth = "200px", separator = c("Format"))
)
}
\arguments{
\item{summarisedResults}{Summary characteristics long table.}

\item{style}{}
}
\value{
New table in gt format
}
\description{
Create a gt table from a summary characteristics object
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

summariseCharacteristics(
  cohort = cdm$cohort1,
  ageGroup = list(c(0, 19), c(20, 39), c(40, 59), c(60, 79), c(80, 150)),
  tableIntersect = list(
    "Visits" = list(
      tableName = "visit_occurrence", value = "count", window = c(-365, 0)
     )
  ),
  cohortIntersect = list(
    "Medications" = list(
      targetCohortTable = "cohort2", value = "flag", window = c(-365, 0)
    )
  ),
  minCellCount = 1
) \%>\%
  gtCharacteristics()
}
}
