% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockPatientProfiles.R
\name{mockPatientProfiles}
\alias{mockPatientProfiles}
\title{It creates a mock database for testing PatientProfiles package}
\usage{
mockPatientProfiles(
  numberIndividuals = 10,
  ...,
  source = "local",
  con = lifecycle::deprecated(),
  writeSchema = lifecycle::deprecated(),
  seed = lifecycle::deprecated()
)
}
\arguments{
\item{numberIndividuals}{Number of individuals to create in the cdm
reference.}

\item{...}{User self defined tables to put in cdm, it can input as many
as the user want.}

\item{source}{Source for the mock cdm, it can either be 'local' or 'duckdb'.}

\item{con}{deprecated.}

\item{writeSchema}{deprecated.}

\item{seed}{deprecated.}
}
\value{
A mock cdm_reference object created following user's specifications.
}
\description{
It creates a mock database for testing PatientProfiles package
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()

}

}
