% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientProfiles-package.R
\docType{package}
\name{PatientProfiles-package}
\alias{PatientProfiles}
\alias{PatientProfiles-package}
\title{PatientProfiles: Identify characteristics of patients in the OMOP Common Data Model}
\description{
Identify demographics of patients in data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model.
}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk}

Authors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk}
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk}
  \item Kim Lopez-Guell \email{kim.lopez@spc.ox.ac.uk}
}

}
\keyword{internal}
