% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Summarise cohort timing}
\usage{
plotCohortOverlap(
  result,
  percentage = TRUE,
  subjects = TRUE,
  facetBy = NULL,
  cohortLabels = "{cohort_name_reference}; {cohort_name_comparator}"
)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{percentage}{If TRUE the overlap plot will display percentages,
otherwise it will present counts.}

\item{subjects}{If TRUE the overlap plot will display number of subjects,
otherwise it will present number of records.}

\item{facetBy}{Names of columns in the result dataframe for faceting the
ggplot object.}

\item{cohortLabels}{A glue expression to identify each plotted cohort
overlap.}
}
\value{
A summarised result
}
\description{
Summarise cohort timing
}
\examples{
\donttest{
}

}
