% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Plot cohort_overlap objects.}
\usage{
plotCohortOverlap(
  result,
  cohortNameReference = NULL,
  cohortNameComparator = NULL,
  strataName = NULL,
  strataLevel = NULL,
  cdmName = NULL,
  variableName = "number subjects",
  facetBy = NULL,
  overlapLabel = "{cohort_name_reference}; {cohort_name_comparator}",
  uniqueCombinations = TRUE
)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{cohortNameReference}{Names of the reference cohorts to include.}

\item{cohortNameComparator}{Names of the comparator cohorts to include.}

\item{strataName}{Names of the strata names to include.}

\item{strataLevel}{Names of the strata levels to include.}

\item{cdmName}{Names of the databases to include.}

\item{variableName}{Names of the variable names to include.}

\item{facetBy}{Names of columns in the cohort_overlap table for faceting the
ggplot object.}

\item{overlapLabel}{A glue expression to identify each plotted cohort
overlap.}

\item{uniqueCombinations}{If TRUE, only unique combinations of reference and
comparator plots will be plotted.}
}
\value{
A ggplot2.
}
\description{
Plot cohort_overlap objects.
}
\examples{
\donttest{
}

}
