% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Plot cohort_overlap objects.}
\usage{
plotCohortOverlap(
  result,
  facetBy = "variable_name",
  overlapLabel = "{cohort_name_reference}; {cohort_name_comparator}",
  uniqueCombinations = TRUE
)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{facetBy}{Names of columns in the cohort_overlap table for faceting the
ggplot object.}

\item{overlapLabel}{A glue expression to identify each plotted cohort
overlap.}

\item{uniqueCombinations}{If TRUE, only unique combinations of reference and
comparator plots will be plotted.}
}
\value{
A ggplot2.
}
\description{
Plot cohort_overlap objects.
}
\examples{
\donttest{
library(PatientProfiles)
cdm <- PatientProfiles::mockPatientProfiles()
overlap <- summariseCohortOverlap(cdm$cohort2)
plotCohortOverlap(overlap)
}

}
