% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSummarised.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Summarise cohort timing}
\usage{
plotCohortOverlap(result, percentage = TRUE, subjects = TRUE, facetBy = NULL)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}

\item{restrictToFirstEntry}{If TRUE only an individual's first entry per
cohort will be considered. If FALSE all entries per individual will be
considered}

\item{timing}{Summary statistics for timing.}
}
\value{
A summarised result
}
\description{
Summarise cohort timing
}
\examples{
\donttest{
cdm_local <- omock::mockCdmReference() |>
  omock::mockPerson(100) |>
  omock::mockObservationPeriod() |>
  omock::mockCohort(numberCohorts = 2)

con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
cdm <- CDMConnector::copy_cdm_to(con = con,
                                 cdm = cdm_local,
                                 schema = "main")

timing <- summariseCohortTiming(cdm$cohort)
}

}
