% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCharacteristics.R
\name{summariseCharacteristics}
\alias{summariseCharacteristics}
\title{Summarise the characteristics of different individuals}
\usage{
summariseCharacteristics(
  table,
  strata = list(),
  variables = list(numericVariables = detectVariables(table, "numeric"), dateVariables =
    detectVariables(table, "date"), binaryVariables = detectVariables(table, "binary"),
    categoricalVariables = detectVariables(table, "categorical")),
  functions = list(numericVariables = c("median", "q25", "q75"), dateVariables =
    c("median", "q25", "q75"), binaryVariables = c("count", "\%"), categoricalVariables =
    c("count", "\%")),
  suppressCellCount = 5,
  bigMark = ",",
  decimalMark = ".",
  significantDecimals = 2
)
}
\arguments{
\item{table}{Table with different records}

\item{strata}{List of the stratifications to be considered.}

\item{variables}{List of the different groups of variables, by default they
are automatically classified.}

\item{functions}{List of functions to be applied to each one of the group of
variables.}

\item{suppressCellCount}{Minimum count of records to report results.}

\item{bigMark}{Big mark delimiter.}

\item{decimalMark}{Decimal separator.}

\item{significantDecimals}{Number of significant decimals reported.}
}
\value{
Table that summarises the characteristics of the individual.
}
\description{
Summarise the characteristics of different individuals
}
\examples{
\donttest{
library(PatientProfiles)
library(dplyr)

cdm <- mockPatientProfiles()
x <- cdm$cohort1 \%>\%
  addDemographics(cdm) \%>\%
  collect()
result <- summariseCharacteristics(x)
}

}
