% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortOverlap.R
\name{summariseCohortOverlap}
\alias{summariseCohortOverlap}
\title{Summarise cohort overlap}
\usage{
summariseCohortOverlap(cohort)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference}
}
\value{
A summarised_result
}
\description{
Summarise cohort overlap
}
\examples{
\donttest{
cdm_local <- omock::mockCdmReference() |>
  omock::mockPerson(100) |>
  omock::mockObservationPeriod() |>
  omock::mockCohort(numberCohorts = 2)

con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
cdm <- CDMConnector::copy_cdm_to(con = con,
                                 cdm = cdm_local,
                                 schema = "main")

overlap <- summariseCohortOverlap(cdm$cohort)
}
}
