% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseLargeScaleCharacteristics.R
\name{summariseLargeScaleCharacteristics}
\alias{summariseLargeScaleCharacteristics}
\title{This function is used to summarise the large scale characteristics of a
cohort table}
\usage{
summariseLargeScaleCharacteristics(
  cohort,
  strata = list(),
  window = list(c(-Inf, -366), c(-365, -31), c(-30, -1), c(0, 0), c(1, 30), c(31, 365),
    c(366, Inf)),
  incidentStandard = NULL,
  overlapStandard = NULL,
  incidentSource = NULL,
  overlapSource = NULL,
  concepSetList = NULL,
  incidentAtc = NULL,
  overlapAtc = NULL,
  incidentIcd10 = NULL,
  overlapIcd10 = NULL,
  incidentConceptSet = NULL,
  overlapConceptSet = NULL,
  minCellCount = 5,
  cdm = attr(cohort, "cdm_reference")
)
}
\arguments{
\item{cohort}{The cohort to characterise.}

\item{strata}{Stratification list.}

\item{window}{Temporal windows that we want to characterize.}

\item{incidentStandard}{Tables to characterise the concept_id.}

\item{overlapStandard}{Tables to characterise the concept_id.}

\item{incidentSource}{Tables to characterise the source_concept_id.}

\item{overlapSource}{Tables to characterise the source_concept_id.}

\item{incidentAtc}{ATC levels to characterise.}

\item{overlapAtc}{ATC levels to characterise.}

\item{incidentIcd10}{ICD10 levels to characterise.}

\item{overlapIcd10}{ICD10 levels to characterise.}

\item{incidentConceptSet}{Concept sets to characterise.}

\item{overlapConceptSet}{Concept sets to characterise.}

\item{minCellCount}{All counts lower than minCellCount will be obscured.}

\item{cdm}{A cdm reference.}
}
\value{
The output of this function is a `ResultSummary` containing the
relevant information.
}
\description{
This function is used to summarise the large scale characteristics of a
cohort table
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

cdm$cohort1 \%>\%
  summariseLargeScaleCharacteristics(
    window = list(c(-180, -1), c(0, 0), c(1, 180)),
    incidentStandard = "condition_occurrence",
    overlapStandard = "drug_expoure"
  )
}

}
