% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseResult.R
\name{suppressCounts}
\alias{suppressCounts}
\title{Function to suppress counts in summarised objects}
\usage{
suppressCounts(result, minCellCount = 5)
}
\arguments{
\item{result}{SummarisedResult object}

\item{minCellCount}{Minimum count of records to report results.}
}
\value{
Table with suppressed counts
}
\description{
Function to suppress counts in summarised objects
}
