% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummary.R
\name{tableCharacteristics}
\alias{tableCharacteristics}
\title{Create a gt table from a summary characteristics object}
\usage{
tableCharacteristics(table)
}
\arguments{
\item{table}{Summary characteristics long table.}
}
\value{
New table in gt format
}
\description{
Create a gt table from a summary characteristics object
}
\examples{
\donttest{
library(PatientProfiles)

cdm <- mockPatientProfiles()
cdm$cohort1 \%>\%
  summariseCharacteristics(
    cdm = cdm, minCellCount = 1, covariates = list(cohort2 = c(-Inf, -1))
  ) \%>\%
  tableCharacteristics()
}

}
