% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableFunctions.R
\name{tableCohortOverlap}
\alias{tableCohortOverlap}
\title{Format a cohort_overlap object into a visual table.}
\usage{
tableCohortOverlap(
  result,
  cohortNameReference = NULL,
  cohortNameComparator = NULL,
  strataName = "overall",
  strataLevel = "overall",
  splitStrata = TRUE,
  cdmName = NULL,
  variableName = c("number records", "number subjects"),
  type = "gt",
  minCellCount = 5,
  .options = list()
)
}
\arguments{
\item{result}{A cohort_overlap object.}

\item{cohortNameReference}{Names of the reference cohorts to include.}

\item{cohortNameComparator}{Names of the comparator cohorts to include.}

\item{strataName}{Names of the strata names to include.}

\item{strataLevel}{Names of the strata levels to include.}

\item{splitStrata}{If TRUE strata name-level columns will be split.}

\item{cdmName}{Name of the databases to include.}

\item{variableName}{Name of the variable names to include.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", or "tibble".}

\item{minCellCount}{Counts below which results will be clouded.}

\item{.options}{named list with additional formatting options.
PatientProfiles::optionsTableCohortOverlap() shows allowed arguments and
their default values.}
}
\value{
A formatted table of the overlap_cohort summarised object.
}
\description{
Format a cohort_overlap object into a visual table.
}
\examples{
\donttest{
}

}
