% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattedTables.R
\name{tableCohortOverlap}
\alias{tableCohortOverlap}
\title{Format a summarised_characteristics object into a visual table.}
\usage{
tableCohortOverlap(
  result,
  type = "gt",
  numberSubjects = TRUE,
  numberRecords = TRUE,
  cdmName = TRUE,
  minCellCount = 5,
  .options = list()
)
}
\arguments{
\item{result}{A summarised_characteristics object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", or "tibble".}

\item{numberSubjects}{Whether to include the number of subjects.}

\item{numberRecords}{Whether to include the number of records}

\item{cdmName}{Whether to include the CDM name.}

\item{minCellCount}{Counts below which results will be clouded.}

\item{.options}{named list with additional formatting options.
PatientProfiles::tableCohortOverlapOptions() shows allowed arguments and
their default values.}
}
\value{
A formatted table of the overlap_cohort summarised object.
}
\description{
Format a summarised_characteristics object into a visual table.
}
\examples{
\donttest{
cdm_local <- omock::mockCdmReference() |>
  omock::mockPerson(100) |>
  omock::mockObservationPeriod() |>
  omock::mockCohort(numberCohorts = 2)

con <- DBI::dbConnect(duckdb::duckdb(), ":memory:")
cdm <- CDMConnector::copy_cdm_to(con = con,
                                 cdm = cdm_local,
                                 schema = "main"
cdm$cohort |>
  summariseCohortOverlap() |>
  tableCohortOverlap()
}

}
