% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableSummary.R
\name{tableSummary}
\alias{tableSummary}
\title{Create a gt table from a summary results objects}
\usage{
tableSummary(table, filterRow = NULL, pivotWider = NULL, hideColumn = NULL)
}
\arguments{
\item{table}{Long table.}

\item{filterRow}{Filter variables.}

\item{pivotWider}{List of columns to compare.}

\item{hideColumn}{Columns to hide.}
}
\value{
New table in gt format
}
\description{
Create a gt table from a summary results objects
}
