% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{variableTypes}
\alias{variableTypes}
\title{Classify the variables between 5 types: "numeric", "categorical", "binary",
"date", or NA.}
\usage{
variableTypes(x)
}
\arguments{
\item{x}{Tibble with different columns.}
}
\value{
Tibble with the variables type and classification
}
\description{
Classify the variables between 5 types: "numeric", "categorical", "binary",
"date", or NA.
}
\examples{
library(PatientProfiles)
library(tibble)
x <- tibble(
  person_id = c(1, 2),
  start_date = as.Date(c("2020-05-02", "2021-11-19")),
  asthma = c(0, 1)
)
variableTypes(x)

}
